Attribute Macro dynec::system

#[system]

Converts a function into a system.

This macro converts the function into a unit struct with the same name that implements system::Spec. The unit struct also derefs to a function pointer, so it is still possible to call the function directly (mainly useful in unit tests) without any change in the signature. However it is not recommended to call this function directly in production code.

Options

Options are applied behind the attribute name in the form #[system(...)]. Multiple options are separated by commas.

name = $expr

Sets the name of the system to $expr. By default, the name is concat!(module_path!(), "::", $function_identifier).

The $expr can read the local and param states of the system directly. Since the debug name is only used for display purposes, it is allowed (although confusing to the programmer) to use mutable states in the name. It is unspecified whether debug messages use the initial debug name or the updated state.

before($expr1, $expr2, ...) and after($expr1, $expr2, ...)

Indicates that the system must be executed before/after all partitions given in the expressions.

Similar to name, the expressions can read local and param states dirctly. However, only the expressions are only resolved once before the first run of the system, so mutating states has no effect on the system schedule.

Parameters

Each parameter of a system function has a special meaning:

Local states

Parameters with the attribute #[dynec(local = xxx)] are “local states”, where xxx is an expression that evaluates to the initial value of the state.

Local states must take the type &T or &mut T, where T is the actual stored state. The mutated state persists for each instance of the system.

Use global states instead if the local state needs to be accessed from multiple systems.

Param states

Parameters with the attribute #[dynec(param)] are “param states”. The user has to pass initial values for param states in the .build() method. Param states behave identically to local states except for different definition location of the initial value.

It is typically used to initialize systems with resources that cannot be created statically (e.g. system canvas resources), or to schedule multiple systems declared from the same function (e.g. working on multiple discriminants of an isotope component).

Global states

Parameters with the attribute #[dynec(global)] are “global states”. Global states are shared scalar data between multiple systems. See Global for more information.

Thread-unsafe (non-Send + Sync) global states must be declared as #[dynec(global(thread_local))] to indicate that the global state can only be accessed from the main thread. As a result, systems that request thread-local global states will only be scheduled on the main thread.

Example

use dynec::system;

#[dynec::global(initial = Title("hello world"))]
struct Title(&'static str);

#[derive(PartialEq, Eq, Hash)]
struct Foo;

#[system(
    name = format!("simulate[one = {}, two = {}]", counter_one, counter_two),
    before(Foo),
)]
fn simulate(
    #[dynec(local = 0)] counter_one: &mut u16,
    #[dynec(param)] counter_two: &mut i64,
    #[dynec(global)] title: &mut Title,
) {
    *counter_one += 1u16;
    *counter_two += 3i64;

    if *counter_two == 5 {
        title.0 = "changed";
    }
}

{
    // We can call the function directly in unit tests.

    let mut counter_one = 0u16;
    let mut counter_two = 2i64;
    let mut title = Title("original");

    simulate(&mut counter_one, &mut counter_two, &mut title);

    assert_eq!(counter_one, 1u16);
    assert_eq!(counter_two, 5i64);
    assert_eq!(title.0, "changed");
}

let spec = simulate.build(7i64);
assert_eq!(system::Spec::debug_name(&spec), "simulate[one = 0, two = 7]");