Module dynec::comp::any

Utilities for dynamic dispatch related to components.

Macros

impl_isotope_init_fn 🔒

impl_isotope_init_fn_accumulate 🔒

impl_simple_init_fn 🔒

impl_simple_init_fn_accumulate 🔒

Structs

Identifier 🔒

Identifies a generic simple or discriminated isotope component type.

IsotopeIniter

Describes how to instantiate a component based on other component types.

Map

A generic TypeMap of owned simple and isotope components.

SimpleIniter

Describes how to instantiate a component based on other component types.

Traits

IsotopeInitFn

A function used for comp::IsotopeInitStrategy::Auto.

SimpleInitFn

A function used for comp::SimpleInitStrategy::Auto.

Type Definitions

DepList

Dependency list.