Trait dynec::comp::discrim::Discrim

pub trait Discrim: Debug + Copy + PartialEq + Eq + Hash + Send + Sync + 'static {
    type FullMap<S>: FullMap<Discrim = Self, Key = Self, Value = S>;

    // Required methods
    fn from_usize(usize: usize) -> Self;
    fn into_usize(self) -> usize;
}

A discriminant value that distinguishes different isotopes of the same component type.

A discriminant should have a one-to-one mapping to usize, which is used to represent the discriminant in type-erased code (such as scheduling). Furthermore, if FullMap is LinearVecMap, SortedVecMap or BoundedVecMap, this usize is used for indexing storages during all-isotopes read/write access. The range of mapped usizes should be bounded to a small number if BoundedVecMap is used.

Required Associated Types

type FullMap<S>: FullMap<Discrim = Self, Key = Self, Value = S>

The data structure that can efficiently access an item based on discriminants from a Set.

Required Methods

fn from_usize(usize: usize) -> Self

Constructs a discriminant from the usize.

The returned value must be consistent and inverse of into_usize.

Can panic if the usize is not supported.

fn into_usize(self) -> usize

Converts the discriminant to a usize.

The returned value must be consistent and inverse of from_usize. discrim1 == discrim2 if and only if discrim1.into_usize() == discrim2.into_usize().