Trait dynec::comp::discrim::FullMap

pub trait FullMap
where
    Self: Mapped<Key = <Self as Mapped>::Discrim> + FromIterator<(Self::Discrim, Self::Value)>,
{
    type MapRef<U>: Mapped<Discrim = Self::Discrim, Key = Self::Key, Value = U>;

    // Required methods
    fn map_ref<U, F>(&self, f: F) -> Self::MapRef<U>
       where F: FnMut(&Self::Value) -> U;
    fn get_by_or_insert<F: FnOnce() -> Self::Value>(
        &mut self,
        discrim: Self::Discrim,
        inserter: F
    ) -> &mut Self::Value;
    fn get_by_or_insert_array<'t, T, Inserter, Transform, const N: usize>(
        &'t mut self,
        discrims: [Self::Discrim; N],
        inserter: Inserter,
        transform: Transform
    ) -> [T; N]
       where T: RefToPtr,
             Inserter: FnMut(Self::Discrim) -> Self::Value,
             Transform: FnMut(Self::Discrim, &'t mut Self::Value) -> T;
}

A data structure to index objects by all known discriminants.

This is only used when storages of all isotopes are read/written in the same accessor (through Components::read_full_isotope_storage, or system::ReadIsotopeFull))]`).

Required Associated Types

type MapRef<U>: Mapped<Discrim = Self::Discrim, Key = Self::Key, Value = U>

Return value of map_ref.

Required Methods

fn map_ref<U, F>(&self, f: F) -> Self::MapRef<U>
where F: FnMut(&Self::Value) -> U,

Maps all values in this map, taking self by shared reference, preserving a similar data structure.

fn get_by_or_insert<F: FnOnce() -> Self::Value>( &mut self, discrim: Self::Discrim, inserter: F ) -> &mut Self::Value

Lazily initializes the entry and returns a mutable reference

fn get_by_or_insert_array<'t, T, Inserter, Transform, const N: usize>( &'t mut self, discrims: [Self::Discrim; N], inserter: Inserter, transform: Transform ) -> [T; N]
where T: RefToPtr, Inserter: FnMut(Self::Discrim) -> Self::Value, Transform: FnMut(Self::Discrim, &'t mut Self::Value) -> T,

Lazily initializes the entry and returns a mutable reference

Object Safety

This trait is not object safe.