Trait dynec::comp::simple::Simple

pub trait Simple<A: Archetype>: Referrer + Send + Sync + Sized + 'static {
    type Storage: Storage<RawEntity = A::RawEntity, Comp = Self>;

    const PRESENCE: SimplePresence;
    const INIT_STRATEGY: SimpleInitStrategy<A>;
    const IS_FINALIZER: bool = false;
}

A simple component has only one instance per entity.

See the module-level documentation for more information.

Required Associated Types

type Storage: Storage<RawEntity = A::RawEntity, Comp = Self>

The storage type used for storing this simple component.

Required Associated Constants

const PRESENCE: SimplePresence

The presence constraint of this component.

const INIT_STRATEGY: SimpleInitStrategy<A>

The initialization strategy for this component.

Provided Associated Constants

const IS_FINALIZER: bool = false

Override this to true if the component is a finalizer.

Finalizer components must be optional. Entities are not removed until all finalizer components have been removed.