Trait dynec::comp::Isotope

pub trait Isotope<A: Archetype>: Referrer + Send + Sync + Sized + 'static {
    type Discrim: Discrim;
    type Storage: Storage<RawEntity = A::RawEntity, Comp = Self>;

    const INIT_STRATEGY: IsotopeInitStrategy<A>;
}

An isotope component may have multiple instances per entity.

See the module-level documentation for more information.

Required Associated Types

type Discrim: Discrim

The discriminant type.

type Storage: Storage<RawEntity = A::RawEntity, Comp = Self>

The storage type used for storing this simple component.

Required Associated Constants

const INIT_STRATEGY: IsotopeInitStrategy<A>

The initialization strategy for this component.