Trait dynec::comp::IsotopeInitFn

pub trait IsotopeInitFn<A: Archetype>: Send + Sync + 'static {
    fn populate(&self, map: &mut Map<A>);
    fn deps(&self) -> DepList<A>;
}

A function used for comp::IsotopeInitStrategy::Auto.

This trait is blanket-implemented for all functions that take up to 32 isotope component parameters and output an iterator of (discriminant, component) tuples.

Required Methods

fn populate(&self, map: &mut Map<A>)

Calls the underlying function, extracting the arguments.

fn deps(&self) -> DepList<A>

Returns the component types required by this function.