Trait dynec::comp::Must

pub trait Must<A: Archetype>: SimpleOrIsotope<A> { }

Marks that a component type is always present.

This trait must only be implemented by components that implement SimpleOrIsotope with SimpleOrIsotope::PRESENCE set to Presence::Required.

Not implementing this trait does not result in any issues except for ergonomic inconvenience when using getters on storages.

Object Safety

This trait is not object safe.