Trait dynec::comp::Simple

pub trait Simple<A: Archetype>: SimpleOrIsotope<A> {
    const IS_FINALIZER: bool = false;
}

A simple component has only one instance per entity.

See the module-level documentation for more information.

Provided Associated Constants

const IS_FINALIZER: bool = false

Override this to true if the component is a finalizer.

Finalizer components must be optional. Entities are not removed until all finalizer components have been removed.