Trait dynec::component::Discrim

pub trait Discrim: Copy {
    fn from_usize(usize: usize) -> Self;
    fn to_usize(self) -> usize;
}

A discriminant value that distinguishes different isotopes of the same component type.

For compact storage, the discriminant should have a one-to-one mapping to the usize type. The usize needs not be a small number; it can be any valid usize as long as it is one-to-one and consistent.

Required Methods

fn from_usize(usize: usize) -> Self

Constructs a discriminant from the usize.

Can panic if the usize is not supported.

fn to_usize(self) -> usize

Converts the discriminant to a usize.