Module dynec::entity

An entity is a single object that owns components.

Entities are reference-counted in debug mode, so they cannot be copied directly.

All components and states referencing entities must override the for_each_entity method.

All strong references to an entity must be dropped before it gets deleted.

Re-exports

pub use ealloc::Ealloc;

Modules

ealloc

Manages entity ID allocation and deallocation.

raw 🔒

referrer 🔒

sealed 🔒

Structs

Entity

A counted reference to an entity.

Generation 🔒

ReferrerArg

The opaque argument passed through Referrer::visit.

UnclonableRef

An unclonable reference to an entity.

Weak

A weak counted reference to an entity.

Traits

Raw

A raw entity ID.

Ref

A trait implemented by all types of entity references.

Referrer

A type that may own entity references (no matter strong or weak).

Derive Macros

Ref

Re-export of dynec::EntityRef.