Module dynec::entity::ealloc

Manages entity ID allocation and deallocation.

Modules

recycling 🔒

sharding 🔒

snapshot 🔒

Structs

BTreeHint

Additional configuration for allocating entities from a BTreeSet recycler.

Map

Contains entity allocators for all archetypes.

MutTakeIter 🔒

Recycling

The default allocator supporting atomically-allocated new IDs and arbitrary recycler.

RecyclingShard

Shard implementation for Recycling.

ShardMap

A map of shards assigned to a single worker thread.

ShardMapEntry 🔒

Snapshot

A snapshot of the allocated entities during offline.

StaticShardAssigner

A shard assigner that never shuffles and always allocates from the same shard. Used for testing only.

ThreadRngShardAssigner

The default shard assigner using rand::thread_rng.

Traits

AnyEalloc 🔒

Ealloc

Manages sharded entity ID allocation and deallocation.

Recycler

A data structure that provides the ability to recycle entity IDs.

Shard

A sharded entity ID allocator.

ShardAssigner

Provides the randomness for shard assignment.

Functions

builder 🔒

iter_gaps 🔒

Type Aliases

AnyBuilder 🔒

BorrowedShard

Return value of ShardMap::borrow.