Trait dynec::entity::ealloc::Ealloc

pub trait Ealloc: 'static {
    type Raw: Raw;
    type AllocHint: Default;
    type Shard: Shard<Raw = Self::Raw, Hint = Self::AllocHint>;

    fn new(num_shards: usize) -> Self;
    fn shards<U, F: Fn(Self::Shard) -> U>(&mut self, vec: &mut Vec<U>, f: F);
    fn allocate(&mut self, hint: Self::AllocHint) -> Self::Raw;
    fn deallocate(&mut self, ids: Vec<Self::Raw>);
}

Manages sharded entity ID allocation and deallocation.

Required Associated Types

type Raw: Raw

The raw entity ID type supported by this allocator.

type AllocHint: Default

The hint type supported by the allocator to fine-tune memory allocation.

type Shard: Shard<Raw = Self::Raw, Hint = Self::AllocHint>

The shard type sent to each worker thread.

Required Methods

fn new(num_shards: usize) -> Self

Initialize a new allocator with num_shards shards.

fn shards<U, F: Fn(Self::Shard) -> U>(&mut self, vec: &mut Vec<U>, f: F)

Returns a vector of shards where each shard references internal states in the allocator.

The length of the result must be num_shards in Ealloc::new. The caller shall shuffle the results returned by this method.

fn allocate(&mut self, hint: Self::AllocHint) -> Self::Raw

Allocate an ID. Can only be used between out-of-cycle.

fn deallocate(&mut self, ids: Vec<Self::Raw>)

Deallocates the IDs, allowing them to be reused in future calls to Shard::allocate.