Trait dynec::entity::ealloc::Shard

pub trait Shard: Send + 'static {
    type Raw: Raw;
    type Hint: Sized;

    fn allocate(&mut self, hint: Self::Hint) -> Self::Raw;
    fn deallocate(&mut self, id: Self::Raw);
    fn deallocate_buffer(self: Box<Self>, buf: &mut Vec<Self::Raw>);
}

A sharded entity ID allocator.

Each worker thread has mutable access to a shard in each cycle. Between cycles, the shards are shuffled to new worker threads.

Required Associated Types

type Raw: Raw

type Hint: Sized

Required Methods

fn allocate(&mut self, hint: Self::Hint) -> Self::Raw

Allocates an ID from the shard.

fn deallocate(&mut self, id: Self::Raw)

fn deallocate_buffer(self: Box<Self>, buf: &mut Vec<Self::Raw>)