Trait dynec::entity::raw::Raw

pub unsafe trait Raw: Sized + Send + Sync + Copy + Debug + Eq + Ord + 'static {
    type Atomic: Atomic<Self>;
    type Range: Iterator<Item = Self>;

    fn new() -> Self::Atomic;
    fn add(self, count: usize) -> Self;
    fn sub(self, other: Self) -> usize;
    fn from_primitive(i: usize) -> Self;
    fn to_primitive(self) -> usize;
    fn range(range: Range<Self>) -> Self::Range;
}

A raw entity ID.

Types implementing this trait are only used in storage internals.

Safety

Violating the Eq and Ord invariants leads to undefined behavior. Furthermore, from_primitive and to_primitive should preserve the equivalence and ordering.

Required Associated Types

type Atomic: Atomic<Self>

The atomic variant of this data type.

type Range: Iterator<Item = Self>

Return value of range.

Required Methods

fn new() -> Self::Atomic

Creates the smallest value of this type in atomic form, used for initialization.

fn add(self, count: usize) -> Self

Equivalent to self + count, does not mutate any values

fn sub(self, other: Self) -> usize

Equivalent to self - other, does not mutate any values. May panic when self < other.

fn from_primitive(i: usize) -> Self

Converts the primitive scalar to the ID. The primitive scalar is guaranteed to be valid, returned from a previous to_primitive call.

fn to_primitive(self) -> usize

Converts the ID to a primitive scalar.

The returned primitive scalar is used for indexing in Vec-based storages, so it should start from a small number.

fn range(range: Range<Self>) -> Self::Range

Iterates over a range.