Module dynec::entity::referrer

Tracks entity references owned by components and globals. See Referrer for more information.

Modules

sealed 🔒

search_single 🔒

std_impl 🔒

Implement Referrer for standard types.

Structs

AsObject

Wraps a trait object for calling Referrer::visit_mut.

NamedBoxIter 🔒

An iterator over Box<dyn Object> that delegates to each object.

NamedIter 🔒

An iterator over T: Object that delegates to each object.

SingleVtable 🔒

Virtual dispatch table to operate referrer functions on single instances, used on global states.

UnnamedIter 🔒

VisitTypeArg

The opaque argument passed to Referrer::visit_type. This type is used to hide the implementation detail from users such that the actual arguments are only visible to the internals.

Traits

Object 🔒

Trait objects that are capable of calling Referrer::visit_mut with specific implementors of VisitMutArg.

Referrer

A type that may own entity references (no matter strong or weak).

VisitMutArg

The trait bound for arguments passed to Referrer::visit_mut.