Trait dynec::entity::Referrer

pub trait Referrer {
    fn visit(&mut self, arg: &mut ReferrerArg<'_>);
}

A type that may own entity references (no matter strong or weak).

The parameters passed in this trait are abstracted by the opaque type ReferrerArg. Implementors should only forward the arg reference to the other implementors, where the actual logic is eventually implemented by owned Entity and Weak fields.

This trait is deliberately not implemented for UnclonableRef, because this trait should only be used in global states and components, whilst UnclonableRef should only be used in systems temporarily.

Required Methods

fn visit(&mut self, arg: &mut ReferrerArg<'_>)

Performs the mapping and increments the counter for each entity.

Each entity reference must be visited only exactly once for each visitor. As a result, Referrer is not implemented for Arc because it may result in visiting the same entity reference multiple times, which will lead to incorrect behaviour.