Macro dynec::zip

zip!() { /* proc-macro */ }

Declares a composite struct that implements IntoZip, Zip and ZipChunked by delegation to all fields and reconstructing the same struct with different types.

All fields accept arbitrary types, similar to a tuple, and are projected to the corresponding storages upon entity iteration.

Example

#![feature(return_position_impl_trait_in_trait)]

dynec::zip! {
    /// This is an example zip struct.
    /// We can document it and apply attributes on it.
    #[allow(dead_code)]
    pub Foo {
        /// This documents the field.
        pub(crate) bar,
        qux,
    }
}