Module dynec::macros

Macros

archetype

Declares archetypes.

comps

Creates a map of components for a given archetype.

zip

Declares a composite struct that implements IntoZip, Zip and ZipChunked by delegation to all fields and reconstructing the same struct with different types.

Attribute Macros

comp

Derives a comp::Simple or comp::Isotope implementation for the applied type. This macro does not modify the input other than stripping attributes.

global

Derives a Global implementation for the applied type. This macro does not modify the input other than stripping attributes.

system

Converts a function into a system.

Derive Macros

Discrim

Derives a Discrim implementation for the type.

EntityRef

Derives a Referrer implementation for the type.