Trait dynec::storage::simple::AnySimpleStorage

pub(crate) trait AnySimpleStorage<A: Archetype>: Send + Sync {
    // Required methods
    fn as_any(&self) -> &(dyn Any + Send + Sync);
    fn as_any_mut(&mut self) -> &mut (dyn Any + Send + Sync);
    fn fill_init_simple(
        &mut self,
        entity: A::RawEntity,
        comp_map: &mut Map<A>,
        dep_getter: DepGetter<'_, A>
    );
    fn has_finalizer(&self, entity: A::RawEntity) -> bool;
    fn clear_entry(&mut self, entity: A::RawEntity);
    fn referrer_dyn<'t>(&'t mut self) -> Box<dyn Object + 't>;
    fn get_any(&self, entity: A::RawEntity) -> Option<&dyn Any>;
}

Required Methods

fn as_any(&self) -> &(dyn Any + Send + Sync)

fn as_any_mut(&mut self) -> &mut (dyn Any + Send + Sync)

fn fill_init_simple( &mut self, entity: A::RawEntity, comp_map: &mut Map<A>, dep_getter: DepGetter<'_, A> )

Fills a simple component of with the initial value.

fn has_finalizer(&self, entity: A::RawEntity) -> bool

Returns true if C::IS_FINALIZER. and the component exists for the given entity.

fn clear_entry(&mut self, entity: A::RawEntity)

Clears the component data for an entity if any.

fn referrer_dyn<'t>(&'t mut self) -> Box<dyn Object + 't>

Returns a referrer::Object implementation that visits all components in this storage.

fn get_any(&self, entity: A::RawEntity) -> Option<&dyn Any>

Returns the value for an entity and return it as an Any.

Due to the poor performance of Any, this method should not be used unless component type elision is necessary.

Implementations

impl<A: Archetype> dyn AnySimpleStorage<A>

pub(crate) fn downcast_ref<C: Simple<A>>(&self) -> &C::Storage

pub(crate) fn downcast_mut<C: Simple<A>>(&mut self) -> &mut C::Storage