Trait dynec::storage::Access

pub trait Access {
    type RawEntity: Raw;
    type Comp: Send + Sync + 'static;
    type IterMut<'u>: Iterator<Item = (Self::RawEntity, &'u mut Self::Comp)> + 'u
       where Self: 'u;

    // Required methods
    fn get_mut(&mut self, entity: Self::RawEntity) -> Option<&mut Self::Comp>;
    fn iter_mut(&mut self) -> Self::IterMut<'_>;
}

Mutable access functions for a storage, generalizing Storage and Partition.

Required Associated Types

type RawEntity: Raw

The type of entity ID used for identification.

type Comp: Send + Sync + 'static

The component type stored.

type IterMut<'u>: Iterator<Item = (Self::RawEntity, &'u mut Self::Comp)> + 'u where Self: 'u

Return value of iter_mut.

Required Methods

fn get_mut(&mut self, entity: Self::RawEntity) -> Option<&mut Self::Comp>

Gets a mutable reference to the component for a specific entity if it is present.

fn iter_mut(&mut self) -> Self::IterMut<'_>

Returns a mutable iterator over the storage, ordered by entity index order.