Trait dynec::storage::Partition

pub trait Partition<'t>: Access + Send + Sync + Sized + 't {
    type ByRef<'u>: Partition<'u, RawEntity = Self::RawEntity, Comp = Self::Comp>
       where Self: 'u;
    type IntoIterMut: Iterator<Item = (Self::RawEntity, &'t mut Self::Comp)>;

    // Required methods
    fn by_ref(&mut self) -> Self::ByRef<'_>;
    fn split_out(&mut self, entity: Self::RawEntity) -> Self;
    fn into_iter_mut(self) -> Self::IntoIterMut;
    fn into_mut(self, entity: Self::RawEntity) -> Option<&'t mut Self::Comp>;
    fn into_many_mut<const N: usize>(
        self,
        entities: [Self::RawEntity; N]
    ) -> Option<[&'t mut Self::Comp; N]>;

    // Provided method
    fn split_at(self, entity: Self::RawEntity) -> (Self, Self) { ... }
}

Borrows a slice of a storage, analogously &'t mut Storage[..].

This trait does not provide set because adding/removing items may cause rebalances in the tree implementation and result in dangling references in other partitions that are not &mut-locked.

Required Associated Types

type ByRef<'u>: Partition<'u, RawEntity = Self::RawEntity, Comp = Self::Comp> where Self: 'u

Return value of by_ref.

type IntoIterMut: Iterator<Item = (Self::RawEntity, &'t mut Self::Comp)>

Return value of into_iter_mut.

Required Methods

fn by_ref(&mut self) -> Self::ByRef<'_>

Re-borrows the partition with reduced lifetime.

This is useful for calling into_iter_mut and split_at, which take self as receiver to preserve the lifetime.

fn split_out(&mut self, entity: Self::RawEntity) -> Self

Splits the partition further into two subpartitions, replacing self with the left partition.

entity must be > 0 and < partition_length, i.e. the expected key ranges of both partitions must be nonempty. (It is allowed to have a nonempty range which does not contain any existing keys)

fn into_iter_mut(self) -> Self::IntoIterMut

Same as iter_mut, but moves the partition object into the iterator.

fn into_mut(self, entity: Self::RawEntity) -> Option<&'t mut Self::Comp>

Same as get_mut, but returns a reference with lifetime 't.

fn into_many_mut<const N: usize>( self, entities: [Self::RawEntity; N] ) -> Option<[&'t mut Self::Comp; N]>

Same as get_many_mut, but returns a reference with lifetime 't.

Provided Methods

fn split_at(self, entity: Self::RawEntity) -> (Self, Self)

Splits the partition further into two subpartitions. entity must be > 0 and < partition_length, i.e. the expected key ranges of both partitions must be nonempty. (It is allowed to have a nonempty range which does not contain any existing keys)

Object Safety

This trait is not object safe.