Trait dynec::storage::PartitionChunked

pub trait PartitionChunked<'t>: Partition<'t> + AccessChunked {
    type IntoIterChunksMut: Iterator<Item = (Self::RawEntity, &'t mut [Self::Comp])>;

    // Required methods
    fn into_chunk_mut(
        self,
        start: Self::RawEntity,
        end: Self::RawEntity
    ) -> Option<&'t mut [Self::Comp]>;
    fn into_iter_chunks_mut(self) -> Self::IntoIterChunksMut;
}

Borrows a slice of a chunked storage, analogously &'t mut Chunked[..].

This trait does not provide set because adding/removing items may cause rebalances in the tree implementation and result in dangling references in other partitions that are not &mut-locked.

Required Associated Types

type IntoIterChunksMut: Iterator<Item = (Self::RawEntity, &'t mut [Self::Comp])>

Return value of into_iter_chunks_mut.

Required Methods

fn into_chunk_mut( self, start: Self::RawEntity, end: Self::RawEntity ) -> Option<&'t mut [Self::Comp]>

Gets a mutable reference to a slice of components, preserving the lifetime 't of this partition object.

fn into_iter_chunks_mut(self) -> Self::IntoIterChunksMut

Returns a mutable iterator over the storage, ordered by entity index order.

Object Safety

This trait is not object safe.