Module dynec::system

Systems are actions performed every frame to manipulate entities and components.

A system can declare mutable or immutable access to archetype-specific components and global states. dynec will schedule systems such that systems have unique access to mutable resources they request. Furthermore, systems can request relative execution order through Partitions. Systems that use thread-unsafe resources (systems that are not Send) are always executed on the main thread.

Modules

spec

Specifies the requirements for a system.

Structs

IsotopeMutMap

Provides mutable access to all isotopes of the same type for an entity.

IsotopeRefMap

Provides immutable access to all isotopes of the same type for an entity.

Spec

Describes an instance of system.

Traits

Partition

A partition is a hashable type constructed by system specifications used to constrain system execution order. Two partition objects are considered equivalent if they have the same type and hash.

ReadIsotope

Provides access to an isotope component in a specific archetype.

ReadSimple

Provides access to a simple component in a specific archetype.

Sendable

A system requests some resources, stores some states of its own and is runnable with the requested resources.

Unsendable

A variant of Sendable that runs on the main thread only, but allows storing Send states and accessing non-Send + Sync global states.

WriteIsotope

Provides access to an isotope component in a specific archetype.

WriteSimple

Provides access to a simple component in a specific archetype.