Struct dynec::system::access::isotope::Isotope

pub struct Isotope<A, C, StorageMapT> {
    storages: StorageMapT,
    _ph: PhantomData<(A, C)>,
}

Accesses multiple storages for the same isotope.

Fields
storages: StorageMapT_ph: PhantomData<(A, C)>
Implementations

impl<A, C, StorageMapT> Isotope<A, C, StorageMapT>

pub(crate) fn new(storages: StorageMapT) -> Self

impl<A, C, KeyT, StorageMapT> Isotope<A, C, StorageMapT>
where A: Archetype, C: Isotope<A>, KeyT: Debug + Copy + 'static, StorageMapT: StorageMap<A, C, Key = KeyT>,

pub fn get(&mut self, entity: impl Ref<Archetype = A>, discrim: KeyT) -> &C
where C: Must<A>,

Retrieves the component for the given entity and discriminant.

This method is infallible for correctly implemented comp::Must, which returns the auto-initialized value for missing components.

pub fn try_get( &mut self, entity: impl Ref<Archetype = A>, key: KeyT ) -> Option<&C>

Returns an immutable reference to the component for the specified entity and discriminant, or the default value for isotopes with a default initializer or None if the component is not present in the entity.

pub fn known_discrims<'t>( &'t self ) -> impl Iterator<Item = <C as Isotope<A>>::Discrim> + 't

Iterates over all known discriminants of the component type.

The yielded discriminants are not in any guaranteed order.

pub fn get_all<'t, E: Ref<Archetype = A>>( &'t self, entity: E ) -> impl Iterator<Item = (<C as Isotope<A>>::Discrim, &'t C)> + 't

Iterates over all known isotopes for a specific entity.

pub fn iter<'t>( &'t mut self, key: KeyT ) -> impl Iterator<Item = (TempRef<'t, A>, &'t C)>

Iterates over all components of a specific discriminant.

Note that the initializer is not called for lazy-initialized isotope components. To avoid confusing behavior, do not use this function if C: comp::Must<A>.

pub fn split<'t, const N: usize>( &'t mut self, keys: [KeyT; N] ) -> [AccessSingle<A, C, impl Deref<Target = <C as SimpleOrIsotope<A>>::Storage> + 't>; N]

Splits the accessor into multiple immutable AccessSingle accessors so that they can be used independently.

impl<A, C, KeyT, StorageMapT> Isotope<A, C, StorageMapT>
where A: Archetype, C: Isotope<A>, KeyT: Debug + Copy + 'static, StorageMapT: PartialStorageMap<A, C, Key = KeyT>,

pub fn get_ref(&self, entity: impl Ref<Archetype = A>, key: KeyT) -> &C
where C: Must<A>,

Retrieves the component for the given entity and discriminant.

Identical to get but does not require a mutable receiver.

pub fn try_get_ref<E: Ref<Archetype = A>>( &self, entity: E, key: KeyT ) -> Option<&C>

Returns an immutable reference to the component for the specified entity and discriminant, or the default value for isotopes with a default initializer or None if the component is not present in the entity.

Identical to try_get but does not require a mutable receiver.

pub fn iter_ref<'t>( &'t self, key: KeyT ) -> impl Iterator<Item = (TempRef<'t, A>, &'t C)>

Iterates over all components of a specific discriminant.

Identical to iter but does not require a mutable receiver.

impl<A, C, KeyT, StorageMapT> Isotope<A, C, StorageMapT>
where A: Archetype, C: Isotope<A>, KeyT: Debug + Copy + 'static, StorageMapT: StorageMapMut<A, C, Key = KeyT>,

pub fn get_mut( &mut self, entity: impl Ref<Archetype = A>, discrim: KeyT ) -> &mut C
where C: Must<A>,

Retrieves the component for the given entity and discriminant.

This method is infallible for correctly implemented comp::Must, which returns the auto-initialized value for missing components.

pub fn try_get_mut( &mut self, entity: impl Ref<Archetype = A>, key: KeyT ) -> Option<&mut C>

Returns a mutable reference to the component for the specified entity and discriminant, automatically initialized with the default initializer if present, or None if the component is unset and has no default initializer.

Note that this method returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use set to add/remove a component.

pub fn set<E: Ref<Archetype = A>>( &mut self, entity: E, key: KeyT, value: Option<C> ) -> Option<C>

Overwrites the component for the specified entity and discriminant.

Passing None to this method removes the component from the entity.

pub fn iter_mut<'t>( &'t mut self, key: KeyT ) -> impl Iterator<Item = (TempRef<'t, A>, &'t mut C)>

Iterates over mutable references to all components of a specific discriminant.

pub fn split_mut<'t, const N: usize>( &'t mut self, keys: [KeyT; N] ) -> [AccessSingle<A, C, impl DerefMut<Target = <C as SimpleOrIsotope<A>>::Storage> + 't>; N]

Splits the accessor into multiple mutable AccessSingle accessors so that they can be used in entity iteration independently.