Trait dynec::system::access::isotope::Get

pub trait Get {
    type Arch: Archetype;
    type Comp: Isotope<Self::Arch>;
    type Key: Debug + Copy + 'static;
    type KnownDiscrims<'t>: Iterator<Item = <Self::Comp as Isotope<Self::Arch>>::Discrim> + 't
       where Self: 't;
    type GetAll<'t, E: Ref<Archetype = Self::Arch>>: Iterator<Item = (<Self::Comp as Isotope<Self::Arch>>::Discrim, &'t Self::Comp)> + 't
       where Self: 't;
    type Iter<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)>
       where Self: 't;
    type Split<'t, const N: usize>: Deref<Target = <Self::Comp as SimpleOrIsotope<Self::Arch>>::Storage> + 't
       where Self: 't;

    // Required methods
    fn get(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>,
        discrim: Self::Key
    ) -> &Self::Comp
       where Self::Comp: Must<Self::Arch>;
    fn try_get(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>,
        key: Self::Key
    ) -> Option<&Self::Comp>;
    fn known_discrims<'t>(&'t self) -> Self::KnownDiscrims<'t>;
    fn get_all<'t, E: Ref<Archetype = Self::Arch>>(
        &'t self,
        entity: E
    ) -> Self::GetAll<'t, E>;
    fn iter<'t>(&'t mut self, key: Self::Key) -> Self::Iter<'t>;
    fn split<'t, const N: usize>(
        &'t mut self,
        keys: [Self::Key; N]
    ) -> [AccessSingle<Self::Arch, Self::Comp, Self::Split<'t, N>>; N];
}

Derived trait for Isotope.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: Isotope<Self::Arch>

The component that this accessor retrieves.

type Key: Debug + Copy + 'static

The key for the discriminant set, Comp::Discrim for full accessors, typically usize for partial accessors

type KnownDiscrims<'t>: Iterator<Item = <Self::Comp as Isotope<Self::Arch>>::Discrim> + 't where Self: 't

Return value for known_discrims

type GetAll<'t, E: Ref<Archetype = Self::Arch>>: Iterator<Item = (<Self::Comp as Isotope<Self::Arch>>::Discrim, &'t Self::Comp)> + 't where Self: 't

Return value for get_all

type Iter<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)> where Self: 't

Return value for iter

type Split<'t, const N: usize>: Deref<Target = <Self::Comp as SimpleOrIsotope<Self::Arch>>::Storage> + 't where Self: 't

Return value for split

Required Methods

fn get( &mut self, entity: impl Ref<Archetype = Self::Arch>, discrim: Self::Key ) -> &Self::Comp
where Self::Comp: Must<Self::Arch>,

Retrieves the component for the given entity and discriminant.

This method is infallible for correctly implemented comp::Must, which returns the auto-initialized value for missing components.

fn try_get( &mut self, entity: impl Ref<Archetype = Self::Arch>, key: Self::Key ) -> Option<&Self::Comp>

Returns an immutable reference to the component for the specified entity and discriminant, or the default value for isotopes with a default initializer or None if the component is not present in the entity.

fn known_discrims<'t>(&'t self) -> Self::KnownDiscrims<'t>

Iterates over all known discriminants of the component type.

The yielded discriminants are not in any guaranteed order.

fn get_all<'t, E: Ref<Archetype = Self::Arch>>( &'t self, entity: E ) -> Self::GetAll<'t, E>

Iterates over all known isotopes for a specific entity.

fn iter<'t>(&'t mut self, key: Self::Key) -> Self::Iter<'t>

Iterates over all components of a specific discriminant.

Note that the initializer is not called for lazy-initialized isotope components. To avoid confusing behavior, do not use this function if C: comp::Must<A>.

fn split<'t, const N: usize>( &'t mut self, keys: [Self::Key; N] ) -> [AccessSingle<Self::Arch, Self::Comp, Self::Split<'t, N>>; N]

Splits the accessor into multiple immutable AccessSingle accessors so that they can be used independently.

Object Safety

This trait is not object safe.