Trait dynec::system::access::isotope::GetMut

pub trait GetMut {
    type Arch: Archetype;
    type Comp: Isotope<Self::Arch>;
    type Key: Debug + Copy + 'static;
    type IterMut<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t mut Self::Comp)>
       where Self: 't;
    type SplitIsotopes<'t, const N: usize>: DerefMut<Target = <Self::Comp as SimpleOrIsotope<Self::Arch>>::Storage> + 't
       where Self: 't;

    // Required methods
    fn get_mut(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>,
        discrim: Self::Key
    ) -> &mut Self::Comp
       where Self::Comp: Must<Self::Arch>;
    fn try_get_mut(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>,
        key: Self::Key
    ) -> Option<&mut Self::Comp>;
    fn set<E: Ref<Archetype = Self::Arch>>(
        &mut self,
        entity: E,
        key: Self::Key,
        value: Option<Self::Comp>
    ) -> Option<Self::Comp>;
    fn iter_mut<'t>(&'t mut self, key: Self::Key) -> Self::IterMut<'t>;
    fn split_isotopes<'t, const N: usize>(
        &'t mut self,
        keys: [Self::Key; N]
    ) -> [AccessSingle<Self::Arch, Self::Comp, Self::SplitIsotopes<'t, N>>; N];
}

Derived trait for Isotope.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: Isotope<Self::Arch>

The component that this accessor retrieves.

type Key: Debug + Copy + 'static

The key for the discriminant set, Comp::Discrim for full accessors, typically usize for partial accessors

type IterMut<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t mut Self::Comp)> where Self: 't

Return value for iter_mut

type SplitIsotopes<'t, const N: usize>: DerefMut<Target = <Self::Comp as SimpleOrIsotope<Self::Arch>>::Storage> + 't where Self: 't

Return value for split_isotopes

Required Methods

fn get_mut( &mut self, entity: impl Ref<Archetype = Self::Arch>, discrim: Self::Key ) -> &mut Self::Compwhere Self::Comp: Must<Self::Arch>,

Retrieves the component for the given entity and discriminant.

This method is infallible for correctly implemented comp::Must, which returns the auto-initialized value for missing components.

fn try_get_mut( &mut self, entity: impl Ref<Archetype = Self::Arch>, key: Self::Key ) -> Option<&mut Self::Comp>

Returns a mutable reference to the component for the specified entity and discriminant, automatically initialized with the default initializer if present, or None if the component is unset and has no default initializer.

Note that this method returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use set to add/remove a component.

fn set<E: Ref<Archetype = Self::Arch>>( &mut self, entity: E, key: Self::Key, value: Option<Self::Comp> ) -> Option<Self::Comp>

Overwrites the component for the specified entity and discriminant.

Passing None to this method removes the component from the entity.

fn iter_mut<'t>(&'t mut self, key: Self::Key) -> Self::IterMut<'t>

Iterates over mutable references to all components of a specific discriminant.

fn split_isotopes<'t, const N: usize>( &'t mut self, keys: [Self::Key; N] ) -> [AccessSingle<Self::Arch, Self::Comp, Self::SplitIsotopes<'t, N>>; N]

Splits the accessor into multiple mutable AccessSingle accessors so that they can be used in entity iteration independently.