Trait dynec::system::access::isotope::GetRef

pub trait GetRef {
    type Arch: Archetype;
    type Comp: Isotope<Self::Arch>;
    type Key: Debug + Copy + 'static;
    type IterRef<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)>
       where Self: 't;

    // Required methods
    fn get_ref(
        &self,
        entity: impl Ref<Archetype = Self::Arch>,
        key: Self::Key
    ) -> &Self::Comp
       where Self::Comp: Must<Self::Arch>;
    fn try_get_ref<E: Ref<Archetype = Self::Arch>>(
        &self,
        entity: E,
        key: Self::Key
    ) -> Option<&Self::Comp>;
    fn iter_ref<'t>(&'t self, key: Self::Key) -> Self::IterRef<'t>;
}

Derived trait for Isotope.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: Isotope<Self::Arch>

The component that this accessor retrieves.

type Key: Debug + Copy + 'static

The key for the discriminant set, Comp::Discrim for full accessors, typically usize for partial accessors

type IterRef<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)> where Self: 't

Return value for iter_ref

Required Methods

fn get_ref( &self, entity: impl Ref<Archetype = Self::Arch>, key: Self::Key ) -> &Self::Comp
where Self::Comp: Must<Self::Arch>,

Retrieves the component for the given entity and discriminant.

Identical to get but does not require a mutable receiver.

fn try_get_ref<E: Ref<Archetype = Self::Arch>>( &self, entity: E, key: Self::Key ) -> Option<&Self::Comp>

Returns an immutable reference to the component for the specified entity and discriminant, or the default value for isotopes with a default initializer or None if the component is not present in the entity.

Identical to try_get but does not require a mutable receiver.

fn iter_ref<'t>(&'t self, key: Self::Key) -> Self::IterRef<'t>

Iterates over all components of a specific discriminant.

Identical to iter but does not require a mutable receiver.

Object Safety

This trait is not object safe.