Trait dynec::system::access::isotope::PartialStorageMap

pub trait PartialStorageMap<A, C>: StorageMap<A, C>
where
    A: Archetype,
    C: Isotope<A>,
{
    // Required method
    fn get_storage_ref(&self, key: Self::Key) -> &C::Storage;
}

Like StorageMap but can access a storage without &mut self.

Only available for partial accessors, because a full accessor needs to mutate its local copy of storage map.

Required Methods

fn get_storage_ref(&self, key: Self::Key) -> &C::Storage

Retrieves a storage by key like get_storage, but without exclusively borrowing the accessor.

Object Safety

This trait is not object safe.