Trait dynec::system::access::isotope::StorageMap

pub trait StorageMap<A, C>where
    A: Archetype,
    C: Isotope<A>,{
    type Key: Debug + Copy + 'static;
    type IterKeys<'t>: Iterator<Item = (Self::Key, C::Discrim)> + 't
       where Self: 't;
    type IterValue: Deref<Target = C::Storage>;
    type IterValues<'t>: Iterator<Item = (Self::Key, C::Discrim, &'t Self::IterValue)> + 't
       where Self: 't;

    // Required methods
    fn get_storage(&mut self, key: Self::Key) -> &C::Storage;
    fn get_storage_many<const N: usize>(
        &mut self,
        keys: [Self::Key; N]
    ) -> [&C::Storage; N];
    fn iter_keys(&self) -> Self::IterKeys<'_>;
    fn iter_values(&self) -> Self::IterValues<'_>;
}

Implements the access pattern for multiple isotope storages.

Required Associated Types

type Key: Debug + Copy + 'static

The key from the user, equivalent to comp::discrim::Set::Key

type IterKeys<'t>: Iterator<Item = (Self::Key, C::Discrim)> + 't where Self: 't

Return value of iter_keys.

type IterValue: Deref<Target = C::Storage>

Storage type yielded by iter_values.

type IterValues<'t>: Iterator<Item = (Self::Key, C::Discrim, &'t Self::IterValue)> + 't where Self: 't

Return value of iter_values.

Required Methods

fn get_storage(&mut self, key: Self::Key) -> &C::Storage

Retrieves a storage by key. Panics if the key is not supported.

For partial accessors, this should return the storage for the discriminant indexed by the key, or panic if the key is out of bounds.

For full accessors, this should return the storage for the given discriminant, or initialize the storage lazily.

fn get_storage_many<const N: usize>( &mut self, keys: [Self::Key; N] ) -> [&C::Storage; N]

Equivalent to calling Self::get_storage for each key.

Duplicate keys are allowed because the return type is immutable. The mutability is only used for lazy initialization.

fn iter_keys(&self) -> Self::IterKeys<'_>

Iterates over all keys currently accessible from this accessor.

For partial accessors, this is the set of keys to the discriminants provided by the user.

For full accessors, this is the set of discriminants that have been initialized.

fn iter_values(&self) -> Self::IterValues<'_>

Iterates over all storages currently accessible from this accessor.

For partial accessors, this is the set of keys to the discriminants provided by the user.

For full accessors, this is the set of discriminants that have been initialized.

Object Safety

This trait is not object safe.