Trait dynec::system::access::isotope::StorageMapMut

pub trait StorageMapMut<A, C>: StorageMap<A, C>where
    A: Archetype,
    C: Isotope<A>,{
    // Required methods
    fn get_storage_mut(&mut self, key: Self::Key) -> &mut C::Storage;
    fn get_storage_mut_many<const N: usize>(
        &mut self,
        keys: [Self::Key; N]
    ) -> [&mut C::Storage; N];
}

Implements the access pattern for multiple isotope storages.

Required Methods

fn get_storage_mut(&mut self, key: Self::Key) -> &mut C::Storage

Retrieves a storage by key. Panics if the key is not supported.

For partial accessors, this should return the storage for the discriminant indexed by the key, or panic if the key is out of bounds.

For full accessors, this should return the storage for the given discriminant, or initialize the storage lazily.

fn get_storage_mut_many<const N: usize>( &mut self, keys: [Self::Key; N] ) -> [&mut C::Storage; N]

Retrieves storages by disjoint keys. Panics if any key is not supported or is equal to another key.