Struct dynec::system::access::single::Single

pub struct Single<A, C, StorageRef> {
    storage: StorageRef,
    _ph: PhantomData<(A, C)>,
}

Access a single component storage, i.e. a simple archetyped component or an isotope archetyped component for a single discriminant.

Fields
storage: StorageRef_ph: PhantomData<(A, C)>
Implementations

impl<A, C, StorageRef> Single<A, C, StorageRef>

pub(crate) fn new(storage: StorageRef) -> Self

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A>, StorageRef: Deref + Sync, StorageRef::Target: Storage<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn try_get(&self, entity: impl Ref<Archetype = A>) -> Option<&C>

Returns an immutable reference to the component for the specified entity, or None if the component is not present in the entity.

pub fn iter<'t>(&'t self) -> impl Iterator<Item = (TempRef<'t, A>, &'t C)> + 't

Iterates over all initialized components in this storage.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageRef: Deref + Sync, StorageRef::Target: Storage<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn get(&self, entity: impl Ref<Archetype = A>) -> &C

Returns an immutable reference to the component for the specified entity.

This function is infallible, assuming comp::Must is only implemented for components with Required presence.

Panics

This function panics if the entity is not fully initialized yet. This happens when an entity is newly created and the cycle hasn’t joined yet.

pub fn par_iter<'t>( &'t self, snapshot: &'t Snapshot<<A as Archetype>::RawEntity> ) -> impl ParallelIterator<Item = (TempRef<'t, A>, &'t C)>

Iterates over chunks of entities in parallel.

This returns a rayon ParallelIterator that processes different chunks of entities

Requires comp::Must<A> because this iterator assumes that existence in snapshot implies existence in storage.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A>, StorageRef: Deref + Sync, StorageRef::Target: Chunked<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn get_chunk(&self, chunk: TempRefChunk<'_, A>) -> &[C]

Returns the chunk of components as a slice.

Panics

This function panics if any component in the chunk is missing. In general, users should not get an entity::TempRefChunk that includes an uninitialized entity, so panic is basically impossible if comp::Must was implemented correctly.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageRef: Deref + Sync, StorageRef::Target: Chunked<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn par_iter_chunks<'t>( &'t self, snapshot: &'t Snapshot<<A as Archetype>::RawEntity> ) -> impl ParallelIterator<Item = (TempRefChunk<'t, A>, &'t [C])>

Iterates over chunks of entities in parallel.

This returns a rayon ParallelIterator that processes different chunks of entities

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A>, StorageRef: DerefMut + Sync, StorageRef::Target: Access<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn try_get_mut(&mut self, entity: impl Ref<Archetype = A>) -> Option<&mut C>

Returns a mutable reference to the component for the specified entity, or None if the component is not present in the entity.

Note that this function returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use set to add/remove a component.

pub fn try_get_many_mut<const N: usize>( &mut self, entities: [impl Ref<Archetype = A>; N] ) -> Option<[&mut C; N]>

Returns mutable references to the components for the specified entities.

Returns None if any component is not present in the entity or if the same entity is passed multiple times.

pub fn iter_mut<'t>( &'t mut self ) -> impl Iterator<Item = (TempRef<'t, A>, &'t mut C)> + 't

Iterates over mutable references to all initialized components in this storage.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageRef: DerefMut + Sync, StorageRef::Target: Access<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn get_mut(&mut self, entity: impl Ref<Archetype = A>) -> &mut C

Returns a mutable reference to the component for the specified entity.

This function is infallible, assuming comp::Must is only implemented for components with Required presence.

Panics

This function panics if the entity is not fully initialized yet. This happens when an entity is newly created and the cycle hasn’t joined yet.

pub fn get_many_mut<const N: usize>( &mut self, entities: [impl Ref<Archetype = A>; N] ) -> [&mut C; N]

Returns mutable references to the component for the specified entities.

Panics

Panics if entities contains duplicate items.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A>, StorageRef: DerefMut + Sync, StorageRef::Target: Storage<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn set( &mut self, entity: impl Ref<Archetype = A>, value: Option<C> ) -> Option<C>

Overwrites the component for the specified entity.

Passing None to this function removes the component from the entity. This leads to a panic for components with comp::Presence::Required presence.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A>, StorageRef: DerefMut + Sync, StorageRef::Target: Storage<RawEntity = A::RawEntity, Comp = C>,

pub fn as_partition( &mut self ) -> Single<A, C, OwnedDeref<<StorageRef::Target as Storage>::Partition<'_>>>

Converts the accessor to a mutably borrowed partition that covers all entities.

The actual splitting partitions can be obtained by calling split_at on the returned value.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageRef: DerefMut + Sync, StorageRef::Target: Storage<RawEntity = <A as Archetype>::RawEntity, Comp = C>,

pub fn par_iter_mut<'t>( &'t mut self, snapshot: &'t Snapshot<<A as Archetype>::RawEntity> ) -> impl ParallelIterator<Item = (TempRef<'t, A>, &'t mut C)>

Iterates over all entities in parallel.

This returns a rayon [ParallelIterator] that processes different entities.

impl<'t, A, C, StorageT> Single<A, C, OwnedDeref<StorageT>>
where A: Archetype, C: SimpleOrIsotope<A>, StorageT: Partition<'t, RawEntity = A::RawEntity, Comp = C>,

pub fn split_at(self, entity: A::RawEntity) -> (Self, Self)

Splits the accessor into two partitions.

The first partition accesses all entities less than entity; the second partition accesses all entities greater than or equal to entity.

pub fn split_out(&mut self, entity: A::RawEntity) -> Self

Splits the accessor into two partitions without moving ownership.

Entities less than entity are retained in self, while entities greater than or equal to entity are accessible through the returned partition.

pub fn try_into_mut(self, entity: impl Ref<Archetype = A>) -> Option<&'t mut C>

Gets the component value of an entity accessible by this partition, preserving the lifetime 't of this partition object.

impl<'t, A, C, StorageT> Single<A, C, OwnedDeref<StorageT>>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageT: Partition<'t, RawEntity = A::RawEntity, Comp = C>,

pub fn into_mut(self, entity: impl Ref<Archetype = A>) -> &'t mut C

Gets the component value of an entity accessible by this partition, preserving the lifetime 't of this partition object.

This function is infallible, assuming comp::Must is only implemented for components with Required presence.

Panics

This function panics if the entity is not fully initialized yet. This happens when an entity is newly created and the cycle hasn’t joined yet.

pub fn into_iter_mut(self) -> impl Iterator<Item = (TempRef<'t, A>, &'t mut C)>

Iterates over mutable references to all initialized components in this partition.

impl<A, C, StorageRef> Single<A, C, StorageRef>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageRef: DerefMut + Sync, StorageRef::Target: Chunked<RawEntity = <A as Archetype>::RawEntity, Comp = C>, for<'u> <StorageRef::Target as Storage>::Partition<'u>: PartitionChunked<'u>,

pub fn get_chunk_mut(&mut self, chunk: TempRefChunk<'_, A>) -> &[C]

Returns the chunk of components as a mutable slice.

Panics

This function panics if any component in the chunk is missing. In general, if comp::Must is implemented correctly, users should not obtain an entity::TempRefChunk that includes an uninitialized entity, so panic is practically impossible.

pub fn par_iter_chunks_mut<'t>( &'t mut self, snapshot: &'t Snapshot<<A as Archetype>::RawEntity> ) -> impl ParallelIterator<Item = (TempRefChunk<'t, A>, &'t mut [C])>

Iterates over all entity chunks in parallel.

This returns a rayon [ParallelIterator] that processes different chunks of entities.

impl<'t, A, C, StorageT> Single<A, C, OwnedDeref<StorageT>>
where A: Archetype, C: SimpleOrIsotope<A> + Must<A>, StorageT: PartitionChunked<'t, RawEntity = A::RawEntity, Comp = C>,

pub fn into_chunk_mut(self, chunk: TempRefChunk<'_, A>) -> &'t mut [C]

Returns the chunk of components as a mutable slice, preserving the lifetime 't of this partition object.

Panics

This function panics if any component in the chunk is missing. In general, if comp::Must is implemented correctly, users should not obtain an entity::TempRefChunk that includes an uninitialized entity, so panic is practically impossible.

pub fn into_iter_chunks_mut( self ) -> impl Iterator<Item = (TempRefChunk<'t, A>, &'t mut [C])>

Iterates over mutable references to all initialized components in this storage.