Trait dynec::system::access::single::Get

pub trait Get {
    type Arch: Archetype;
    type Comp: SimpleOrIsotope<Self::Arch>;
    type Iter<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)> + 't
       where Self: 't;

    // Required methods
    fn try_get(
        &self,
        entity: impl Ref<Archetype = Self::Arch>
    ) -> Option<&Self::Comp>;
    fn iter<'t>(&'t self) -> Self::Iter<'t>;
}

Derived trait for Single.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: SimpleOrIsotope<Self::Arch>

The component that this accessor retrieves.

type Iter<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)> + 't where Self: 't

Return value for iter

Required Methods

fn try_get( &self, entity: impl Ref<Archetype = Self::Arch> ) -> Option<&Self::Comp>

Returns an immutable reference to the component for the specified entity, or None if the component is not present in the entity.

fn iter<'t>(&'t self) -> Self::Iter<'t>

Iterates over all initialized components in this storage.

Object Safety

This trait is not object safe.