Trait dynec::system::access::single::GetChunked

pub trait GetChunked {
    type Arch: Archetype;
    type Comp: SimpleOrIsotope<Self::Arch>;

    // Required method
    fn get_chunk(&self, chunk: TempRefChunk<'_, Self::Arch>) -> &[Self::Comp];
}

Derived trait for Single.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: SimpleOrIsotope<Self::Arch>

The component that this accessor retrieves.

Required Methods

fn get_chunk(&self, chunk: TempRefChunk<'_, Self::Arch>) -> &[Self::Comp]

Returns the chunk of components as a slice.

Panics

This function panics if any component in the chunk is missing. In general, users should not get an entity::TempRefChunk that includes an uninitialized entity, so panic is basically impossible if comp::Must was implemented correctly.

Object Safety

This trait is not object safe.