Trait dynec::system::access::single::GetMut

pub trait GetMut {
    type Arch: Archetype;
    type Comp: SimpleOrIsotope<Self::Arch>;
    type IterMut<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t mut Self::Comp)> + 't
       where Self: 't;

    // Required methods
    fn try_get_mut(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>
    ) -> Option<&mut Self::Comp>;
    fn try_get_many_mut<const N: usize>(
        &mut self,
        entities: [impl Ref<Archetype = Self::Arch>; N]
    ) -> Option<[&mut Self::Comp; N]>;
    fn iter_mut<'t>(&'t mut self) -> Self::IterMut<'t>;
}

Derived trait for Single.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: SimpleOrIsotope<Self::Arch>

The component that this accessor retrieves.

type IterMut<'t>: Iterator<Item = (TempRef<'t, Self::Arch>, &'t mut Self::Comp)> + 't where Self: 't

Return value for iter_mut

Required Methods

fn try_get_mut( &mut self, entity: impl Ref<Archetype = Self::Arch> ) -> Option<&mut Self::Comp>

Returns a mutable reference to the component for the specified entity, or None if the component is not present in the entity.

Note that this function returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use set to add/remove a component.

fn try_get_many_mut<const N: usize>( &mut self, entities: [impl Ref<Archetype = Self::Arch>; N] ) -> Option<[&mut Self::Comp; N]>

Returns mutable references to the components for the specified entities.

Returns None if any component is not present in the entity or if the same entity is passed multiple times.

fn iter_mut<'t>(&'t mut self) -> Self::IterMut<'t>

Iterates over mutable references to all initialized components in this storage.

Object Safety

This trait is not object safe.