Trait dynec::system::access::single::MustGet

pub trait MustGet {
    type Arch: Archetype;
    type Comp: SimpleOrIsotope<Self::Arch> + Must<Self::Arch>;
    type ParIter<'t>: ParallelIterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)>
       where Self: 't;

    // Required methods
    fn get(&self, entity: impl Ref<Archetype = Self::Arch>) -> &Self::Comp;
    fn par_iter<'t>(
        &'t self,
        snapshot: &'t Snapshot<<Self::Arch as Archetype>::RawEntity>
    ) -> Self::ParIter<'t>;
}

Derived trait for Single.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: SimpleOrIsotope<Self::Arch> + Must<Self::Arch>

The component that this accessor retrieves.

type ParIter<'t>: ParallelIterator<Item = (TempRef<'t, Self::Arch>, &'t Self::Comp)> where Self: 't

Return value for par_iter

Required Methods

fn get(&self, entity: impl Ref<Archetype = Self::Arch>) -> &Self::Comp

Returns an immutable reference to the component for the specified entity.

This function is infallible, assuming comp::Must is only implemented for components with Required presence.

Panics

This function panics if the entity is not fully initialized yet. This happens when an entity is newly created and the cycle hasn’t joined yet.

fn par_iter<'t>( &'t self, snapshot: &'t Snapshot<<Self::Arch as Archetype>::RawEntity> ) -> Self::ParIter<'t>

Iterates over chunks of entities in parallel.

This returns a rayon ParallelIterator that processes different chunks of entities

Requires comp::Must<A> because this iterator assumes that existence in snapshot implies existence in storage.