Trait dynec::system::access::single::MustGetMut

pub trait MustGetMut {
    type Arch: Archetype;
    type Comp: SimpleOrIsotope<Self::Arch> + Must<Self::Arch>;

    // Required methods
    fn get_mut(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>
    ) -> &mut Self::Comp;
    fn get_many_mut<const N: usize>(
        &mut self,
        entities: [impl Ref<Archetype = Self::Arch>; N]
    ) -> [&mut Self::Comp; N];
}

Derived trait for Single.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: SimpleOrIsotope<Self::Arch> + Must<Self::Arch>

The component that this accessor retrieves.

Required Methods

fn get_mut( &mut self, entity: impl Ref<Archetype = Self::Arch> ) -> &mut Self::Comp

Returns a mutable reference to the component for the specified entity.

This function is infallible, assuming comp::Must is only implemented for components with Required presence.

Panics

This function panics if the entity is not fully initialized yet. This happens when an entity is newly created and the cycle hasn’t joined yet.

fn get_many_mut<const N: usize>( &mut self, entities: [impl Ref<Archetype = Self::Arch>; N] ) -> [&mut Self::Comp; N]

Returns mutable references to the component for the specified entities.

Panics

Panics if entities contains duplicate items.

Object Safety

This trait is not object safe.