Trait dynec::system::access::single::Set

pub trait Set {
    type Arch: Archetype;
    type Comp: SimpleOrIsotope<Self::Arch>;

    // Required method
    fn set(
        &mut self,
        entity: impl Ref<Archetype = Self::Arch>,
        value: Option<Self::Comp>
    ) -> Option<Self::Comp>;
}

Derived trait for Single.

Required Associated Types

type Arch: Archetype

The archetype that this accessor retrieves for.

type Comp: SimpleOrIsotope<Self::Arch>

The component that this accessor retrieves.

Required Methods

fn set( &mut self, entity: impl Ref<Archetype = Self::Arch>, value: Option<Self::Comp> ) -> Option<Self::Comp>

Overwrites the component for the specified entity.

Passing None to this function removes the component from the entity. This leads to a panic for components with comp::Presence::Required presence.

Object Safety

This trait is not object safe.