Module dynec::system::accessor

Component accessor APIs to be used from systems.

Modules

tuple_impl 🔒

Structs

MustReadAccessor

Return value of Read::access.

MustReadChunkSimpleAccessor

MustWriteAccessor

Return value of Write::access_mut.

MustWriteChunkSimpleAccessor

TryReadAccessor

Return value of Read::try_access.

TryWriteAccessor

Return value of Write::try_access_mut.

Traits

Accessor

An accessor that can be used in an entity iteration.

Chunked

An accessor that can be used in chunked entity iteration.

Read

Generalizes ReadSimple and ReadIsotope for a specific discriminant (through ReadIsotope::split).

ReadChunk

Extends Read with chunk reading ability for storages that support chunked access.

ReadIsotope

Provides access to an isotope component in a specific archetype.

ReadSimple

Provides access to a simple component in a specific archetype.

Write

Generalizes WriteSimple and WriteIsotope for a specific discriminant (through WriteIsotope::split_mut).

WriteChunk

Extends Write with chunk writing ability for storages that support chunked access.

WriteIsotope

Provides access to an isotope component in a specific archetype.

WriteSimple

Provides access to a simple component in a specific archetype.