Trait dynec::system::accessor::WriteIsotope

pub trait WriteIsotope<A: Archetype, C: Isotope<A>, K = <C as Isotope<A>>::Discrim>: ReadIsotope<A, C, K>where
    K: Debug + Copy + 'static,{
    type IterMut<'t>: Iterator<Item = (TempRef<'t, A>, &'t mut C)>
    where
        Self: 't;
    type SplitMut<'t>: Write<A, C> + 't
    where
        Self: 't;

    fn try_get_mut<E: Ref<Archetype = A>>(
        &mut self,
        entity: E,
        discrim: K
    ) -> Option<&mut C>;
    fn set<E: Ref<Archetype = A>>(
        &mut self,
        entity: E,
        discrim: K,
        value: Option<C>
    ) -> Option<C>;
    fn iter_mut(&mut self, discrim: K) -> Self::IterMut<'_>;
    fn split_mut<const N: usize>(
        &mut self,
        keys: [K; N]
    ) -> [Self::SplitMut<'_>; N];
}

Provides access to an isotope component in a specific archetype.

Required Associated Types

type IterMut<'t>: Iterator<Item = (TempRef<'t, A>, &'t mut C)>
where
Self: 't

Return value of iter_mut.

type SplitMut<'t>: Write<A, C> + 't
where
Self: 't

Return value of split.

Required Methods

fn try_get_mut<E: Ref<Archetype = A>>(
&mut self,
entity: E,
discrim: K
) -> Option<&mut C>

Returns a mutable reference to the component for the specified entity and discriminant, automatically initialized with the default initializer if present, or None if the component is unset and has no default initializer.

Note that this method returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use WriteIsotope::set to add/remove a component.

fn set<E: Ref<Archetype = A>>(
&mut self,
entity: E,
discrim: K,
value: Option<C>
) -> Option<C>

Overwrites the component for the specified entity and discriminant.

Passing None to this method removes the component from the entity.

fn iter_mut(&mut self, discrim: K) -> Self::IterMut<'_>

Iterates over mutable references to all components of a specific discriminant.

fn split_mut<const N: usize>(&mut self, keys: [K; N]) -> [Self::SplitMut<'_>; N]

Splits the accessor into multiple Write implementors so that they can be used with [with_mut] independently.