Module dynec::system::iter

Iterate over entities of an archetype.

While individual accessors also provide functions like AccessSingle::iter, functions in EntityIterator use the entity indices from the entity allocator snapshot directly, enabling better performance with chunk partitioning.

Modules

tuple_impls 🔒

Structs

EntityIterator

Allows iterating all entities of an archetype.

MustMissingResln

Automatically unwraps storage results.

Read

IntoZip::IntoZip for read-only accessors.

Try

Wrap accessor references with Try to indicate that the result should be an Option.

Write

IntoZip::IntoZip for mutable accessors.

ZipIter 🔒

Enums

TryMissingResln

Returns None if component is missing.

Traits

IntoZip

Values that can be used as a Zip in EntityIterator, similar to IntoIterator for iterators.

MissingResln

Determines how to resolve the case of a missing Result.

Zip

Multiple single accessors zipped together, to be used with EntityIterator::entities_with.

ZipChunked

Zip accessors with the additional condition that all underlying storages support chunked access, to be used with EntityIterator::chunks_with.