Trait dynec::system::partition::Partition

pub trait Partition: Sealed + Send + Sync + 'static {
    // Required methods
    fn describe(&self, f: &mut Formatter<'_>) -> Result;
    fn compute_hash(&self) -> u64;
    fn equals(&self, other: &dyn Partition) -> bool;
    fn as_any(&self) -> &dyn Any;
}

A partition is a hashable type constructed by system specifications used to constrain system execution order. Two partition objects are considered equivalent if they have the same type and hash.

Systems can declare an anterior or posterior dependency on a partition. If multiple systems specify a dependency for an equivalent partition, it is guaranteed that all anterior systems will finish executing before any posterior system starts executing, effectively creating a “partition” between the anterior and posterior systems.

Required Methods

fn describe(&self, f: &mut Formatter<'_>) -> Result

Describes the partition as fmt::Debug.

fn compute_hash(&self) -> u64

Computes the hash of this component.

fn equals(&self, other: &dyn Partition) -> bool

Checks whether two parttions are equivalent.

fn as_any(&self) -> &dyn Any

Converts the object to an Any.