Trait dynec::system::rw::ReadSimple

pub trait ReadSimple<A: Archetype, C: Simple<A>>: Read<A, C> {
    fn access_chunk(&self) -> MustReadChunkSimple<'_, A, C>;
}

Provides access to a simple component in a specific archetype.

Required Methods

fn access_chunk(&self) -> MustReadChunkSimple<'_, A, C>

Returns a Chunked accessor that can be used in EntityIterator to provide chunked iteration to an entity.