Trait dynec::system::rw::Write

pub trait Write<A: Archetype, C: 'static>: Read<A, C> + Mut<A, C> {
    fn set<E: Ref<Archetype = A>>(
        &mut self,
        entity: E,
        value: Option<C>
    ) -> Option<C>;

    fn get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> &mut C
    where
        C: Must<A>,
    { ... }
    fn access_mut(&mut self) -> MustWrite<A, C, &mut Self>
    where
        C: Must<A>,
    { ... }
    fn try_access_mut(&mut self) -> TryWrite<A, C, &mut Self> { ... }
}

Generalizes WriteSimple and WriteIsotope for a specific discriminant (through WriteIsotope::split_isotopes).

Required Methods

fn set<E: Ref<Archetype = A>>(&mut self, entity: E, value: Option<C>) -> Option<C>

Overwrites the component for the specified entity.

Passing None to this method removes the component from the entity. This leads to a panic for components with comp::SimplePresence::Required presence.

Provided Methods

fn get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> &mut Cwhere
C: Must<A>,

Returns a mutable reference to the component for the specified entity.

This method is infallible, assuming comp::Must is only implemented for components with Required presence.

fn access_mut(&mut self) -> MustWrite<A, C, &mut Self>where
C: Must<A>,

Returns an Accessor implementor that yields &C for each entity.

fn try_access_mut(&mut self) -> TryWrite<A, C, &mut Self>

Returns an Accessor implementor that yields Option<&C> for each entity.