Trait dynec::system::Descriptor

pub trait Descriptor {
    fn get_spec(&self) -> Spec;
    fn visit_type(&self, arg: &mut VisitTypeArg<'_>);
    fn visit_mut(&mut self) -> AsObject<'_>;

    fn state_maybe_uninit(&self) -> Vec<TypeId>  { ... }
}

Describes a system.

Required Methods

fn get_spec(&self) -> Spec

Describes the resource requirements of this system.

The method is only called when the system was initially scheduled, but it should return a consistent value.

fn visit_type(&self, arg: &mut VisitTypeArg<'_>)

Delegates to Referrer::visit_type for the states of this system.

fn visit_mut(&mut self) -> AsObject<'_>

Executes the mutable visitor

Provided Methods

fn state_maybe_uninit(&self) -> Vec<TypeId>

By default, systems with states that strongly reference entities always execute before systems that create those entities to ensure the entities initialized. If references to uninitialized entities are tolerable, include the TypeId of the archetype in the returned vector.