Trait dynec::system::MutChunk

pub trait MutChunk<A: Archetype, C: 'static> {
    // Required method
    fn get_chunk_mut(&mut self, chunk: TempRefChunk<'_, A>) -> &mut [C]
       where C: Must<A>;
}

Extends Mut with chunk writing ability for storages that support chunked access.

Required Methods

fn get_chunk_mut(&mut self, chunk: TempRefChunk<'_, A>) -> &mut [C]where C: Must<A>,

Returns the chunk of components as a mutable slice. Typically called from an accessor.

Panics

This method panics if any component in the chunk is missing. In general, if comp::Must is implemented correctly, users should not obtain an entity::TempRefChunk that includes an uninitialized entity, so panic is practically impossible.