Trait dynec::system::MutFull

pub trait MutFull<A: Archetype, C: 'static>: Mut<A, C> {
    type Partition<'t>: MutPartition<'t, A, C>
       where Self: 't;
    type ParIterMut<'t>: ParallelIterator<Item = (TempRef<'t, A>, &'t mut C)>
       where Self: 't,
             C: Must<A>;

    // Required methods
    fn as_partition(&mut self) -> Self::Partition<'_>;
    fn par_iter_mut<'t>(
        &'t mut self,
        snapshot: &'t Snapshot<A::RawEntity>
    ) -> Self::ParIterMut<'t>
       where C: Must<A>;
}

A Mut accessor that supports all entities, in contrast to MutPartition.

Required Associated Types

type Partition<'t>: MutPartition<'t, A, C> where Self: 't

The partitioned type for this accessor.

type ParIterMut<'t>: ParallelIterator<Item = (TempRef<'t, A>, &'t mut C)> where Self: 't, C: Must<A>

Return value of par_iter_mut.

Required Methods

fn as_partition(&mut self) -> Self::Partition<'_>

Converts the accessor to a MutPartition that covers all entities.

The actual splitting partitions can be obtained by calling split_at on the returned value.

fn par_iter_mut<'t>( &'t mut self, snapshot: &'t Snapshot<A::RawEntity> ) -> Self::ParIterMut<'t>where C: Must<A>,

Iterates over all entities in parallel.

This returns a rayon ParallelIterator that processes different entities.