Trait dynec::system::MutFullChunk

pub trait MutFullChunk<A: Archetype, C: 'static>: MutChunk<A, C> {
    type Partition<'t>: MutPartitionChunk<'t, A, C>
       where Self: 't;
    type ParIterChunksMut<'t>: ParallelIterator<Item = (TempRefChunk<'t, A>, &'t mut [C])>
       where Self: 't,
             C: Must<A>;

    // Required methods
    fn as_partition_chunk(&mut self) -> Self::Partition<'_>;
    fn par_iter_chunks_mut<'t>(
        &'t mut self,
        snapshot: &'t Snapshot<A::RawEntity>
    ) -> Self::ParIterChunksMut<'t>
       where C: Must<A>;
}

A MutChunk accessor that supports all entities, in contrast to MutPartitionChunk.

Required Associated Types

type Partition<'t>: MutPartitionChunk<'t, A, C> where Self: 't

The partitioned type for this accessor.

type ParIterChunksMut<'t>: ParallelIterator<Item = (TempRefChunk<'t, A>, &'t mut [C])> where Self: 't, C: Must<A>

Return value of par_iter_chunks_mut.

Required Methods

fn as_partition_chunk(&mut self) -> Self::Partition<'_>

Converts the accessor to a MutPartitionChunk that covers all entities.

The actual splitting partitions can be obtained by calling split_at on the returned value.

fn par_iter_chunks_mut<'t>( &'t mut self, snapshot: &'t Snapshot<A::RawEntity> ) -> Self::ParIterChunksMut<'t>where C: Must<A>,

Iterates over all entity chunks in parallel.

This returns a rayon ParallelIterator that processes different chunks of entities.