Trait dynec::system::MutPartition

pub trait MutPartition<'t, A: Archetype, C: 'static>: Mut<A, C> + Send + Sized {
    type IntoIterMut: Iterator<Item = (TempRef<'t, A>, &'t mut C)>;

    // Required methods
    fn split_at<E: Ref<Archetype = A>>(self, entity: E) -> (Self, Self);
    fn into_iter_mut(self) -> Self::IntoIterMut;
}

A Mut accessor that can be split into two halves.

Required Associated Types

type IntoIterMut: Iterator<Item = (TempRef<'t, A>, &'t mut C)>

Return value of into_iter_mut.

Required Methods

fn split_at<E: Ref<Archetype = A>>(self, entity: E) -> (Self, Self)

Splits the accessor into two partitions.

The first partition accesses all entities less than entity; the second partition accesses all entities greater than or equal to entity.

fn into_iter_mut(self) -> Self::IntoIterMut

Iterates over mutable references to all initialized components in this storage.