Trait dynec::system::ReadChunk

pub trait ReadChunk<A: Archetype, C: 'static> {
    type ParIterChunks<'t>: ParallelIterator<Item = (TempRefChunk<'t, A>, &'t [C])>
       where Self: 't;

    // Required methods
    fn get_chunk(&self, chunk: TempRefChunk<'_, A>) -> &[C];
    fn par_iter_chunks<'t>(
        &'t self,
        snapshot: &'t Snapshot<A::RawEntity>
    ) -> Self::ParIterChunks<'t>;
}

Extends Read with chunk reading ability for storages that support chunked access.

Required Associated Types

type ParIterChunks<'t>: ParallelIterator<Item = (TempRefChunk<'t, A>, &'t [C])> where Self: 't

Return value of par_iter_chunks.

Required Methods

fn get_chunk(&self, chunk: TempRefChunk<'_, A>) -> &[C]

Returns the chunk of components as a slice.

Panics

This method panics if any component in the chunk is missing. In general, users should not get an entity::TempRefChunk that includes an uninitialized entity, so panic is basically impossible if comp::Must was implemented correctly.

fn par_iter_chunks<'t>( &'t self, snapshot: &'t Snapshot<A::RawEntity> ) -> Self::ParIterChunks<'t>

Iterates over chunks of entities in parallel.

This returns a rayon ParallelIterator that processes different chunks of entities