Trait dynec::system::ReadIsotope

pub trait ReadIsotope<A: Archetype, C: Isotope<A>, K = <C as Isotope<A>>::Discrim>where
    K: Debug + Copy + 'static,{
    type GetAll<'t>: Iterator<Item = (<C as Isotope<A>>::Discrim, &'t C)> + 't
    where
        Self: 't;
    type Iter<'t>: Iterator<Item = (TempRef<'t, A>, &'t C)>
    where
        Self: 't;
    type Split<'t>: Read<A, C> + 't
    where
        Self: 't;

    fn try_get<E: Ref<Archetype = A>>(&self, entity: E, discrim: K) -> Option<&C>;
    fn get_all<E: Ref<Archetype = A>>(&self, entity: E) -> Self::GetAll<'_>;
    fn iter(&self, discrim: K) -> Self::Iter<'_>;
    fn split<const N: usize>(&self, keys: [K; N]) -> [Self::Split<'_>; N];

    fn get<E: Ref<Archetype = A>>(&self, entity: E, discrim: K) -> &C
    where
        C: Must<A>,
    { ... }
}

Provides access to an isotope component in a specific archetype.

K is the type used to index the discriminant. For partial

Required Associated Types

type GetAll<'t>: Iterator<Item = (<C as Isotope<A>>::Discrim, &'t C)> + 't
where
Self: 't

Return value of get_all.

type Iter<'t>: Iterator<Item = (TempRef<'t, A>, &'t C)>
where
Self: 't

Return value of iter.

type Split<'t>: Read<A, C> + 't
where
Self: 't

Return value of split.

Required Methods

fn try_get<E: Ref<Archetype = A>>(&self, entity: E, discrim: K) -> Option<&C>

Returns an immutable reference to the component for the specified entity and discriminant, or the default value for isotopes with a default initializer or None if the component is not present in the entity.

fn get_all<E: Ref<Archetype = A>>(&self, entity: E) -> Self::GetAll<'_>

Iterates over all isotopes of the component type for the given entity.

The yielded discriminants are not in any guaranteed order.

fn iter(&self, discrim: K) -> Self::Iter<'_>

Iterates over all components of a specific discriminant.

Note that the initializer is not called for lazy-initialized isotope components. To avoid confusing behavior, do not use this function if C: comp::Must<A>.

fn split<const N: usize>(&self, keys: [K; N]) -> [Self::Split<'_>; N]

Splits the accessor into multiple Read implementors so that they can be used independently.

Provided Methods

fn get<E: Ref<Archetype = A>>(&self, entity: E, discrim: K) -> &Cwhere
C: Must<A>,

Retrieves the component for the given entity and discriminant.

This method is infallible for correctly implemented comp::Must, which returns the auto-initialized value for missing components.