Trait dynec::system::ReadSimple

pub trait ReadSimple<A: Archetype, C: Simple<A>> {
    fn try_get<E: Ref<Archetype = A>>(&self, entity: E) -> Option<&C>;

    fn get<E: Ref<Archetype = A>>(&self, entity: E) -> &C
    where
        C: Must<A>,
    { ... }
}

Provides access to a simple component in a specific archetype.

Required Methods

fn try_get<E: Ref<Archetype = A>>(&self, entity: E) -> Option<&C>

Returns an immutable reference to the component for the specified entity, or None if the component is not present in the entity.

Provided Methods

fn get<E: Ref<Archetype = A>>(&self, entity: E) -> &C where
C: Must<A>,

Returns an immutable reference to the component for the specified entity.

This method is infallible, assuming comp::Must is only implemented for components with comp::SimplePresence::Required presence.