Trait dynec::system::Sendable

pub trait Sendable: Send {
    fn get_spec(&self) -> Spec;
    fn run(&mut self, globals: &SyncGlobals, components: &Components);
}

A system requests some resources, stores some states of its own and is runnable with the requested resources.

There may be multiple instances of the same implementor type. This is meaningful as they may have different states.

Required Methods

fn get_spec(&self) -> Spec

Describes this instance of system.

The method is only called when the system was initially scheduled, but it should return a consistent value.

fn run(&mut self, globals: &SyncGlobals, components: &Components)

Runs the system.