Trait dynec::system::Unsendable

pub trait Unsendable: Descriptor {
    fn run(
        &mut self,
        sync_globals: &SyncGlobals,
        unsync_globals: &mut UnsyncGlobals,
        components: &Components,
        ealloc_shard_map: &mut ShardMap,
        offline_shard: &mut BufferShard
    );
    fn as_descriptor_mut(&mut self) -> &mut dyn Descriptor;
}

A variant of Sendable that runs on the main thread only, but allows storing Send states and accessing non-Send + Sync global states.

Required Methods

fn run(
&mut self,
sync_globals: &SyncGlobals,
unsync_globals: &mut UnsyncGlobals,
components: &Components,
ealloc_shard_map: &mut ShardMap,
offline_shard: &mut BufferShard
)

Runs the system.

fn as_descriptor_mut(&mut self) -> &mut dyn Descriptor

Returns self upcast to Descriptor as a trait object.