Trait dynec::system::Write

pub trait Write<A: Archetype, C: 'static>: Read<A, C> {
    type IterMut<'t>: Iterator<Item = (TempRef<'t, A>, &'t mut C)>
    where
        Self: 't;

    fn try_get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> Option<&mut C>;
    fn set<E: Ref<Archetype = A>>(
        &mut self,
        entity: E,
        value: Option<C>
    ) -> Option<C>;
    fn iter_mut(&mut self) -> Self::IterMut<'_>;

    fn get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> &mut C
    where
        C: Must<A>,
    { ... }
    fn access_mut(&mut self) -> MustWriteAccessor<A, C, &mut Self>
    where
        C: Must<A>,
    { ... }
    fn try_access_mut(&mut self) -> TryWriteAccessor<A, C, &mut Self> { ... }
}

Generalizes WriteSimple and WriteIsotope for a specific discriminant (through WriteIsotope::split_mut).

Required Associated Types

type IterMut<'t>: Iterator<Item = (TempRef<'t, A>, &'t mut C)>
where
Self: 't

Return value of iter_mut.

Required Methods

fn try_get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> Option<&mut C>

Returns a mutable reference to the component for the specified entity, or None if the component is not present in the entity.

Note that this method returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use set to add/remove a component.

fn set<E: Ref<Archetype = A>>(&mut self, entity: E, value: Option<C>) -> Option<C>

Overwrites the component for the specified entity.

Passing None to this method removes the component from the entity. This leads to a panic for components with comp::SimplePresence::Required presence.

fn iter_mut(&mut self) -> Self::IterMut<'_>

Iterates over mutable references to all initialized components in this storage.

Provided Methods

fn get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> &mut Cwhere
C: Must<A>,

Returns a mutable reference to the component for the specified entity.

This method is infallible, assuming comp::Must is only implemented for components with Required presence.

fn access_mut(&mut self) -> MustWriteAccessor<A, C, &mut Self>where
C: Must<A>,

Returns an Accessor implementor that yields &C for each entity.

fn try_access_mut(&mut self) -> TryWriteAccessor<A, C, &mut Self>

Returns an Accessor implementor that yields Option<&C> for each entity.