Trait dynec::system::WriteChunk

pub trait WriteChunk<A: Archetype, C: 'static> {
    fn get_chunk_mut(&mut self, chunk: TempRefChunk<'_, A>) -> &mut [C] 
    where
        C: Must<A>;
}

Extends Write with chunk writing ability for storages that support chunked access.

Required Methods

fn get_chunk_mut(&mut self, chunk: TempRefChunk<'_, A>) -> &mut [C] where
C: Must<A>,

Returns the chunk of components as a mutable slice. Typically called from an accessor.

Panics

This method panics if any component in the chunk is missing. In general, users should not get an entity::TempRefChunk that includes an uninitialized entity, so panic is basically impossible if comp::Must was implemented correctly.