Type Alias dynec::system::WriteIsotopeFull

pub type WriteIsotopeFull<'t, A, C> = AccessIsotope<A, C, Storages<'t, A, C>>;

Provides access to an isotope component in a specific archetype.

Getters require a mutable receiver to allow lazy initialization of new discriminants. Consider splitting accessors, which returns a system::AccessSingle with a shared receiver. If it can be asserted that no uninitialized discriminants will be encountered, use with known_discrims.

Aliased Type

struct WriteIsotopeFull<'t, A, C> { /* private fields */ }