Module dynec::test_util

Modules

anti_semaphore 🔒

clock 🔒

event_tracer 🔒

globals 🔒

isotope_comps 🔒

simple_comps 🔒

Structs

Aggregator

A generic global state with an initializer.

AntiSemaphore

A synchronization util that blocks until sufficiently many threads are waiting concurrently.

Clock

An emulated clock that supports ticking.

EventTracer

Records event and ensures that they are in the correct order.

InitialEntities

An entity-referencing global state.

IsoNoInit

Does not have auto init

IsoWithInit

Has auto init

Simple1OptionalNoDepNoInit

optional, non-init, depless

Simple2OptionalDepends1

optional, depends on Simple1

Simple3OptionalDepends12

optional, depends on Simple1 + Simple2

Simple4Depends12

optional, depends on Simple1 + Simple2

Simple5RequiredNoInit

required, non-init

Simple6RequiredWithInitNoDeps

required, auto-init, depless

Simple7WithFinalizerNoinit

non-init, has finalizers

SimpleN

a generic component

StrongRefIsotope

An isotope component with a strong reference to TestArch.

StrongRefSimple

A simple component with a strong reference to TestArch.

TestDiscrim1

A test discriminant.

TestDiscrim2

An alternative test discriminant.

use_all_bare

A dummy system used for registering all non-entity-referencing test components.

use_comp_n

A dummy system with minimally simple dependencies.

Enums

TestArch

The default test archetype.

Traits

Tick

Functions

init 🔒