Trait dynec::tracer::Tracer

pub trait Tracer: Sync {
    type CycleContext;
    type PrepareEallocShardsContext;
    type FlushEallocContext;
    type RunSendableContext;
    type RunUnsendableContext;

    // Required methods
    fn start_cycle(&self) -> Self::CycleContext;
    fn end_cycle(&self, arg: Self::CycleContext);
    fn start_prepare_ealloc_shards(&self) -> Self::PrepareEallocShardsContext;
    fn end_prepare_ealloc_shards(&self, arg: Self::PrepareEallocShardsContext);
    fn start_flush_ealloc(
        &self,
        archetype: DbgTypeId
    ) -> Self::FlushEallocContext;
    fn end_flush_ealloc(
        &self,
        arg: Self::FlushEallocContext,
        archetype: DbgTypeId
    );
    fn steal_return_complete(&self, thread: Thread);
    fn steal_return_pending(&self, thread: Thread);
    fn mark_runnable(&self, node: Node);
    fn unmark_runnable(&self, node: Node);
    fn complete_system(&self, node: Node, remaining: usize);
    fn start_run_sendable(
        &self,
        thread: Thread,
        node: Node,
        debug_name: &str,
        system: &mut dyn Sendable
    ) -> Self::RunSendableContext;
    fn end_run_sendable(
        &self,
        context: Self::RunSendableContext,
        thread: Thread,
        node: Node,
        debug_name: &str,
        system: &mut dyn Sendable
    );
    fn start_run_unsendable(
        &self,
        thread: Thread,
        node: Node,
        debug_name: &str,
        system: &mut dyn Unsendable
    ) -> Self::RunUnsendableContext;
    fn end_run_unsendable(
        &self,
        context: Self::RunUnsendableContext,
        thread: Thread,
        node: Node,
        debug_name: &str,
        system: &mut dyn Unsendable
    );
    fn partition(&self, node: Node, partition: &dyn Partition);

}

A handler that receives scheduling-related events in dynec.

New implementations should use #[tracer] to auto implement required types and methods for future compatibility.

Required Associated Types

type CycleContext

Context from start_cycle to end_cycle.

type PrepareEallocShardsContext

Context from start_prepare_ealloc_shards to end_prepare_ealloc_shards.

type FlushEallocContext

Context from start_flush_ealloc to end_flush_ealloc.

type RunSendableContext

Context from start_run_sendable to end_run_sendable.

type RunUnsendableContext

Context from start_run_unsendable to end_run_unsendable.

Required Methods

fn start_cycle(&self) -> Self::CycleContext

A cycle starts.

fn end_cycle(&self, arg: Self::CycleContext)

A cycle ends.

fn start_prepare_ealloc_shards(&self) -> Self::PrepareEallocShardsContext

The executor starts preparing ealloc shards for each worker thread.

fn end_prepare_ealloc_shards(&self, arg: Self::PrepareEallocShardsContext)

The executor has partitioned ealloc into different worker threads.

fn start_flush_ealloc(&self, archetype: DbgTypeId) -> Self::FlushEallocContext

The executor starts preparing ealloc shards for each worker thread.

fn end_flush_ealloc(&self, arg: Self::FlushEallocContext, archetype: DbgTypeId)

The executor has partitioned ealloc into different worker threads.

fn steal_return_complete(&self, thread: Thread)

A thread tries to steal a task, but all tasks have started.

fn steal_return_pending(&self, thread: Thread)

A thread tries to steal a task, but no tasks are in the runnable pool.

fn mark_runnable(&self, node: Node)

A node is marked as runnable because all blockers have been removed.

fn unmark_runnable(&self, node: Node)

A node is unmarked as runnable because an exclusive node has been stolen.

fn complete_system(&self, node: Node, remaining: usize)

A system has completed. Also passes the number of remaining nodes.

fn start_run_sendable( &self, thread: Thread, node: Node, debug_name: &str, system: &mut dyn Sendable ) -> Self::RunSendableContext

A thread-safe system starts running.

fn end_run_sendable( &self, context: Self::RunSendableContext, thread: Thread, node: Node, debug_name: &str, system: &mut dyn Sendable )

A thread-safe system stops running.

fn start_run_unsendable( &self, thread: Thread, node: Node, debug_name: &str, system: &mut dyn Unsendable ) -> Self::RunUnsendableContext

A thread-unsafe system starts running.

fn end_run_unsendable( &self, context: Self::RunUnsendableContext, thread: Thread, node: Node, debug_name: &str, system: &mut dyn Unsendable )

A thread-unsafe system stops running.

fn partition(&self, node: Node, partition: &dyn Partition)

A partition completes.