Trait dynec::world::accessor::MutStorageAccessor

trait MutStorageAccessor<A, C, S, M>where
    A: Archetype,
    C: Isotope<A>,
    S: Deref<Target = C::Storage>,
    M: Mapped<Discrim = C::Discrim>,{
    fn get_storage<'t>(
        &mut self,
        key: M::Key,
        storages: &'t mut M
    ) -> &'t mut C::Storage
    where
        S: 't;
    fn get_storage_multi<'t, const N: usize>(
        &mut self,
        keys: [M::Key; N],
        storages: &'t mut M
    ) -> [&'t mut C::Storage; N]
    where
        S: 't;
}

Determines the strategy to fetch a storage from the storage map. The two implementors from write_full_isotope_storage and write_partial_isotope_storage would either lazily create or panic on missing storage.

This trait is not symmetric to StorageMapProcessorRef because the equivalent API may lead to Polonius compile errors.

Required Methods

fn get_storage<'t>(
&mut self,
key: M::Key,
storages: &'t mut M
) -> &'t mut C::Storagewhere
S: 't,

fn get_storage_multi<'t, const N: usize>(
&mut self,
keys: [M::Key; N],
storages: &'t mut M
) -> [&'t mut C::Storage; N]where
S: 't,