Type Alias dynec::world::rw::isotope::read::partial::ReadIsotopePartial

pub type ReadIsotopePartial<A, C, DiscrimSet = Vec<<C as Isotope<A>>::Discrim>> = AccessIsotope<A, C, Storages<A, C, DiscrimSet>>;

Provides access to an isotope component in a specific archetype.

DiscrimSetKey is the type used to index the discriminant, depending on the type of discriminant set given. For vec and array discriminant sets, this should be usize. For partial isotope access, K is usually usize. For full isotope access, K is the discriminant type.

To share the same API as ReadIsotopeFull, immutable getters still require &mut self, but there are *_ref variants for these functions that just require &self.

Aliased Type

struct ReadIsotopePartial<A, C, DiscrimSet = Vec<<C as Isotope<A>>::Discrim>> {
    storages: Storages<A, C, DiscrimSet>,
    _ph: PhantomData<(A, C)>,
}

Fields
storages: Storages<A, C, DiscrimSet>_ph: PhantomData<(A, C)>