Module dynec::world::state::ouroboros_impl_owning_mapped_rw_lock_read_guard

Encapsulates implementation details for a self-referencing struct. This module is only visible when using –document-private-items.

Structs

BorrowedFields 🔒

A struct for holding immutable references to all tail and immutably borrowed fields in an instance of OwningMappedRwLockReadGuard.

BorrowedMutFields 🔒

A struct for holding mutable references to all tail fields in an instance of OwningMappedRwLockReadGuard.

Heads 🔒

A struct which contains only the head fields of OwningMappedRwLockReadGuard.

OwningMappedRwLockReadGuard 🔒

The self-referencing struct.

OwningMappedRwLockReadGuardAsyncBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

OwningMappedRwLockReadGuardAsyncSendBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

OwningMappedRwLockReadGuardAsyncSendTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of makin your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

OwningMappedRwLockReadGuardAsyncTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of makin your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

OwningMappedRwLockReadGuardBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of making your code both easier to refactor and more readable. Call build() to construct the actual struct. The fields of this struct should be used as follows:

OwningMappedRwLockReadGuardTryBuilder 🔒

A more verbose but stable way to construct self-referencing structs. It is comparable to using StructName { field1: value1, field2: value2 } rather than StructName::new(value1, value2). This has the dual benefit of makin your code both easier to refactor and more readable. Call try_build() or try_build_or_recover() to construct the actual struct. The fields of this struct should be used as follows:

Functions

check_if_okay_according_to_checkers 🔒

type_asserts 🔒