Module dynec::world::storage

A storage is the data structure where components of the same type for all entities are stored.

Structs

ChunkMut

The iterator item of Storage::iter_chunks_mut.

ChunkRef

The iterator item of Storage::iter_chunks.

Tree

A storage based on BTreeMap.

Vec

The basic storage indexed by entity IDs directly.

Traits

Chunked

Provides chunked access capabilities, i.e. the storage can always return a slice for contiguous present components.

Storage

A storage for storing component data.