Trait dynec::world::storage::Chunked

pub unsafe trait Chunked: Storage {
    fn get_chunk(
        &self,
        start: Self::RawEntity,
        end: Self::RawEntity
    ) -> Option<&[Self::Comp]>;
    fn get_chunk_mut(
        &mut self,
        start: Self::RawEntity,
        end: Self::RawEntity
    ) -> Option<&mut [Self::Comp]>;
}

Provides chunked access capabilities, i.e. the storage can always return a slice for contiguous present components.

Safety

Implementors of this trait must ensure that get_chunk and get_chunk_mut are consistent, and non-overlapping ranges map to non-overlapping slices. In other words, for any a, b: Self::RawEntity where a < b, get_chunk(a, b) and get_chunk_mut(a, b) return the same slice (only differing by mutability), and for any c, d: Self::RawEntity where b <= c c < d, get_chunk(a, b) must not alias get_chunk(c, d).

Required Methods

fn get_chunk(
&self,
start: Self::RawEntity,
end: Self::RawEntity
) -> Option<&[Self::Comp]>

Gets a shared reference to a slice of components.

Returns None if any of the components in the range is missing.

Panics if start > end.

fn get_chunk_mut(
&mut self,
start: Self::RawEntity,
end: Self::RawEntity
) -> Option<&mut [Self::Comp]>

Gets a mutable reference to a slice of components.

Returns None if any of the components in the range is missing.

Panics if start > end.