Trait dynec::world::storage::Storage

pub trait Storage: Default + Send + Sync + 'static {
    type RawEntity: Raw;
    type Comp;

    fn get(&self, id: Self::RawEntity) -> Option<&Self::Comp>;
    fn get_mut(&mut self, id: Self::RawEntity) -> Option<&mut Self::Comp>;
    fn set(
        &mut self, 
        id: Self::RawEntity, 
        value: Option<Self::Comp>
    ) -> Option<Self::Comp>;
    fn iter(&self) -> Box<dyn Iterator<Item = (Self::RawEntity, &Self::Comp)>>;
    fn iter_mut(
        &mut self
    ) -> Box<dyn Iterator<Item = (Self::RawEntity, &mut Self::Comp)>>;
}

A storage for storing component data.

Required Associated Types

type RawEntity: Raw

The type of entity ID used for identification.

type Comp

The component type stored.

Required Methods

fn get(&self, id: Self::RawEntity) -> Option<&Self::Comp>

Gets a shared reference to the component for a specific entity if it is present.

fn get_mut(&mut self, id: Self::RawEntity) -> Option<&mut Self::Comp>

Gets a mutable reference to the component for a specific entity if it is present.

fn set(
&mut self,
id: Self::RawEntity,
value: Option<Self::Comp>
) -> Option<Self::Comp>

Sets or removes the component for a specific entity, returning the original value if it was present.

fn iter(&self) -> Box<dyn Iterator<Item = (Self::RawEntity, &Self::Comp)>>

Returns an immutable iterator over the storage, ordered by entity index order.

fn iter_mut(
&mut self
) -> Box<dyn Iterator<Item = (Self::RawEntity, &mut Self::Comp)>>

Returns a mutable iterator over the storage, ordered by entity index order.