Module dynec::world::tracer

Exposes testing, profiling and tracing capabilities.

Macros

define_tracer 🔒

Defines the Tracer trait and implements the Log and Aggregate types.

impl_tuple 🔒

impl_tuple_accumulate 🔒

Structs

Aggregate

Groups multiple tracers into a tuple and dispatches each call to them in serial.

Log

A tracer that logs all events.

Noop

An empty tracer.

RefPartitionWrapper 🔒

Enums

Thread

The thread ID for a system executor.

Traits

Tracer

A tracer used for recording the events throughout an execution cycle.