Attribute Macro dynec::system

#[system]

Converts a function into a system.

This macro converts the function into a unit struct with the same name that implements system::Spec. The unit struct also derefs to a function pointer, so it is still possible to call the function directly (mainly useful in unit tests) without any change in the signature. However it is not recommended to call this function directly in production code.

Options

Options are applied behind the attribute name in the form #[system(...)]. Multiple options are separated by commas.

name = $expr

Sets the name of the system to $expr. By default, the name is concat!(module_path!(), "::", $function_identifier).

The $expr can read the local and param states of the system directly. Since the debug name is only used for display purposes, it is allowed (although confusing to the programmer) to use mutable states in the name. It is unspecified whether debug messages use the initial debug name or the updated state.

before($expr1, $expr2, ...) and after($expr1, $expr2, ...)

Indicates that the system must be executed before/after all partitions given in the expressions.

Similar to name, the expressions can read local and param states dirctly. However, only the expressions are only resolved once before the first run of the system, so mutating states has no effect on the system schedule.

Parameters

Each parameter of a system function has a special meaning:

Local states

Parameters with the attribute #[dynec(local = xxx)] are “local states”, where xxx is an expression that evaluates to the initial value of the state.

Local states must take the type &T or &mut T, where T is the actual stored state. The mutated state persists for each instance of the system.

Use global states instead if the local state needs to be accessed from multiple systems.

Param states

Parameters with the attribute #[dynec(param)] are “param states”. The user has to pass initial values for param states in the .build() method. Param states behave identically to local states except for different definition location of the initial value.

It is typically used to initialize systems with resources that cannot be created statically (e.g. system canvas resources), or to schedule multiple systems declared from the same function (e.g. working on multiple discriminants of an isotope component).

Global states

Parameters with the attribute #[dynec(global)] are “global states”. Global states are shared scalar data between multiple systems. See Global for more information.

Thread-unsafe (non-Send + Sync) global states must be declared as #[dynec(global(thread_local))] to indicate that the global state can only be accessed from the main thread. As a result, systems that request thread-local global states will only be scheduled on the main thread.

Simple components

Parameters with the attribute #[dynec(simple(arch = Type1, comp = Type2))] are “simple components”. Alternatively, parameters that do not have a #[dynec] attribute but with a type impl ReadSimple<Type1, Type2>/impl WriteSimple<Type1, Type2> are also simple components (where Simple can be qualified by any path, e.g. system::Simple). This requests access to a simple component of type Type2 from entities of the archetype Type1, exposed through a type that implements system::ReadSimple or system::WriteSimple.

With the attribute, only read access is permitted by default. To allow write access, add mut to the simple option list with the first syntax. No information is inferred from the type if #[dync(simple)]

Isotope components

Parameters with the attribute #[dynec(isotope(arch = Type1, comp = Type2))] are “isotope components”. Alternatively, parameters that do not have a #[dynec] attribute but with a type Isotope<Type1, &Type2> are also isotope components. This requests access to an isotope component of type Type2 from entities of the archetype Type1, exposed through a type that implements system::ReadIsotope or system::WriteIsotope.

Only read access is permitted by default. To allow write access, add mut to the isotope option list with the first syntax, or change the component reference to &mut with the second syntax.

By default, all discriminants of the isotope component are requested. To request only particular discriminants, add discrim = expr to the isotope option list with the first syntax, or add the #[dynec(isotope(discrim = expr))] with the second syntax, where expr is a value that implements IntoIterator<Item = Type2::Discrim> Note that arch, comp and mut must not be used if the second syntax is intended.

Similar to name, the expressions can read local and param states dirctly. However, only the expressions are only resolved once before the first run of the system, so mutating states has no effect on the system schedule.

Example

use dynec::system;

#[dynec::global(initial = Title("hello world"))]
struct Title(&'static str);

#[derive(Debug, PartialEq, Eq, Hash)]
struct Foo;

dynec::archetype!(Player);

#[dynec::comp(of = Player)]
struct PositionX(f32);
#[dynec::comp(of = Player)]
struct PositionY(f32);

#[dynec::comp(of = Player)]
struct Direction(f32, f32);

#[derive(Debug, Clone, Copy)]
struct SkillId(usize);
impl dynec::comp::Discrim for SkillId {
    fn from_usize(id: usize) -> Self { Self(id) }
    fn into_usize(self) -> usize { self.0 }
}

#[dynec::comp(of = Player, isotope = SkillId, init = || SkillLevel(1))]
struct SkillLevel(u8);

#[system(
    name = format!("simulate[counter = {}, skill_id = {:?}]", counter, skill_id),
    before(Foo),
)]
fn simulate(
    #[dynec(local = 0)] counter: &mut u16,
    #[dynec(param)] &skill_id: &SkillId,
    #[dynec(global)] title: &mut Title,
    x: impl system::WriteSimple<Player, PositionX>,
    y: impl system::WriteSimple<Player, PositionY>,
    dir: impl system::ReadSimple<Player, Direction>,
    #[dynec(isotope(discrim = [skill_id]))] skill: impl system::ReadIsotope<Player, SkillLevel>,
) {
    *counter += 1;

    if *counter == 1 {
        title.0 = "changed";
    }
}

let system = simulate.build(SkillId(3));
assert_eq!(
    system::Sendable::get_spec(&system).debug_name.as_str(),
    "simulate[counter = 0, skill_id = SkillId(3)]"
);

{
    // We can also call the function directly in unit tests.

    let mut counter = 0;
    let mut title = Title("original");

    let mut world = dynec::system_test! {
        simulate.build(SkillId(2));
        _: Player = (
            PositionX(0.0),
            PositionY(0.0),
            Direction(0.5, 0.5),
        );
        _: Player = (
            PositionX(0.5),
            PositionY(0.5),
            Direction(0.5, 0.5),
        );
    };

    /*
    simulate::call(
        &mut counter,
        &SkillId(2),
        &mut title,
        todo!("component access logic"),
        todo!("component access logic"),
        todo!("component access logic"),
        todo!("component access logic"),
    );

    assert_eq!(counter, 1);
    assert_eq!(title.0, "changed");
    */
}