Module dynec::comp::discrim

Discriminants distinguish different isotopes of the same component type.

Structs

ArrayMap

Implements the requirements of Discrim::FullMap with O(1) search complexity, using the into_usize() value as the index.

BoundedVecMap

Implements the requirements of Discrim::FullMap with O(1) search complexity, using the into_usize() value as the index.

LinearVecMap

Implements the requirements of Discrim::FullMap with O(n) search complexity.

SortedVecMap

Implements the requirements of Discrim::FullMap with O(log n) search complexity.

Traits

Discrim

A discriminant value that distinguishes different isotopes of the same component type.

FullMap

A data structure to index objects by all known discriminants.

Mapped

A data structure derived from a discriminant set that can efficiently access an item using Key, which is based on the shape of the set.

Set

A set of discriminants, used for specifying partial access in #[system].

Functions

split_mut_array 🔒 ⚠

Splits an object to multiple mutable references and returns them in an array.