Trait dynec::comp::discrim::Mapped

pub trait Mapped {
    type Discrim: Discrim;
    type Key: Debug + Copy + 'static;
    type Value;
    type Iter<'t>: Iterator<Item = (Self::Discrim, &'t Self::Value)> + 't
    where
        Self: 't;
    type IterMut<'t>: Iterator<Item = (Self::Discrim, &'t mut Self::Value)> + 't
    where
        Self: 't;

    fn get_discrim(&self, key: Self::Key) -> Option<Self::Discrim>;
    fn get_by(&self, key: Self::Key) -> Option<&Self::Value>;
    fn get_mut_by(&mut self, key: Self::Key) -> Option<&mut Self::Value>;
    fn get_mut_array_by<U, TransformFn, OnMissing, const N: usize>(
        &mut self,
        keys: [Self::Key; N],
        transform: TransformFn,
        on_missing: OnMissing
    ) -> [&mut U; N]
    where
        TransformFn: FnMut(&mut Self::Value) -> &mut U,
        OnMissing: FnMut(Self::Key) -> !;
    fn iter_values(&self) -> Self::Iter<'_>;
    fn iter_values_mut(&mut self) -> Self::IterMut<'_>;
}

A data structure derived from a discriminant set that can efficiently access an item using Key, which is based on the shape of the set.

This type is also used for the collection of isotope storages when all discriminants are selected.

Required Associated Types

type Discrim: Discrim

The discriminant type.

type Key: Debug + Copy + 'static

The type used for indexing data.

type Value

The value type stored in this data structure.

type Iter<'t>: Iterator<Item = (Self::Discrim, &'t Self::Value)> + 't
where
Self: 't

Return value of iter_values.

type IterMut<'t>: Iterator<Item = (Self::Discrim, &'t mut Self::Value)> + 't
where
Self: 't

Return value of iter_values_mut.

Required Methods

fn get_discrim(&self, key: Self::Key) -> Option<Self::Discrim>

Gets the discriminant value associated with this key.

fn get_by(&self, key: Self::Key) -> Option<&Self::Value>

Gets a shared reference to an element.

fn get_mut_by(&mut self, key: Self::Key) -> Option<&mut Self::Value>

Executes functions with mutable reference to an entry.

fn get_mut_array_by<U, TransformFn, OnMissing, const N: usize>(
&mut self,
keys: [Self::Key; N],
transform: TransformFn,
on_missing: OnMissing
) -> [&mut U; N]where
TransformFn: FnMut(&mut Self::Value) -> &mut U,
OnMissing: FnMut(Self::Key) -> !,

fn iter_values(&self) -> Self::Iter<'_>

Iterates over the values in this set with the discriminant.

fn iter_values_mut(&mut self) -> Self::IterMut<'_>

Iterates over the values in this set with the discriminant.