Trait dynec::comp::SimpleOrIsotope

pub trait SimpleOrIsotope<A: Archetype>: Referrer + Send + Sync + Sized + 'static {
    type Storage: Storage<RawEntity = A::RawEntity, Comp = Self>;

    const PRESENCE: Presence;
    const INIT_STRATEGY: InitStrategy<A, Self>;
}

The common items for a simple or isotope component.

Required Associated Types

type Storage: Storage<RawEntity = A::RawEntity, Comp = Self>

The storage type used for storing this simple component.

Required Associated Constants

const PRESENCE: Presence

The presence constraint of this component.

const INIT_STRATEGY: InitStrategy<A, Self>

The initialization strategy for this component.

Object Safety

This trait is not object safe.