Module dynec::entity

An entity is a single object that owns components.

Entities are reference-counted in debug mode, so they cannot be copied directly.

All components and states referencing entities must override the for_each_entity method.

All strong references to an entity must be dropped before it gets deleted.

Re-exports

pub use ealloc::Ealloc;

pub use generation::Generation;

pub use referrer::Referrer;

Modules

deletion

Manages entity deletion logic.

ealloc

Manages entity ID allocation and deallocation.

generation

Tracks the number of times an entity ID is allocated, used for distinguishment of dangling weak references.

maybe 🔒

raw 🔒

rctrack 🔒

Last owner of reference-counted entity references, used for identifying strong reference leaks.

referrer

Tracks entity references owned by components and globals. See Referrer for more information.

sealed 🔒

Structs

Entity

A strong reference to an entity.

TempRef

A temporary, non-'static reference to an entity.

TempRefChunk

A chunk of continuous TempRefs.

Weak

A weak counted reference to an entity.

Traits

Raw

A raw entity ID.

Ref

A trait implemented by all types of entity references.

Derive Macros

Ref

Re-export of dynec::EntityRef.