Module dynec::entity::ealloc

Manages entity ID allocation and deallocation.

Structs

BTreeHint

Additional configuration for allocating entities from a BTreeSet recycler.

Map 🔒

MutTakeIter 🔒

Recycling

The default allocator supporting atomically-allocated new IDs and arbitrary recycler.

RecyclingShard

Shard implementation for Recycling.

ShardMap

A map of shards assigned to a single worker thread.

StaticShardAssigner

A shard assigner that never shuffles and always allocates from the same shard. Used for testing only.

ThreadRngShardAssigner

The default shard assigner using rand::thread_rng.

Traits

AnyEalloc 🔒

AnyShard 🔒

Ealloc

Manages sharded entity ID allocation and deallocation.

Recycler

A data structure that provides the ability to recycle entity IDs.

Shard

A sharded entity ID allocator.

ShardAssigner

Provides the randomness for shard assignment.

Functions

builder 🔒

distribute_sorted 🔒

iter_gaps 🔒

Type Definitions

AnyBuilder 🔒

MutableShards 🔒