Trait dynec::entity::ealloc::Ealloc

pub trait Ealloc: 'static {
    type Raw: Raw;
    type AllocHint: Default;
    type Shard: Shard<Raw = Self::Raw, Hint = Self::AllocHint>;

    // Required methods
    fn new(num_shards: usize) -> Self;
    fn shards<U, F: Fn(Self::Shard) -> U>(
        &mut self,
        vec: &mut Vec<U>,
        transform: F
    );
    fn snapshot(&self) -> Snapshot<Self::Raw>;
    fn allocate(&mut self, hint: Self::AllocHint) -> Self::Raw;
    fn queue_deallocate(&mut self, id: Self::Raw);
    fn flush(&mut self);
    fn mark_need_flush(&mut self);
    fn flush_if_marked(&mut self);
}

Manages sharded entity ID allocation and deallocation.

Required Associated Types

type Raw: Raw

The raw entity ID type supported by this allocator.

type AllocHint: Default

The hint type supported by the allocator to fine-tune memory allocation.

type Shard: Shard<Raw = Self::Raw, Hint = Self::AllocHint>

The shard type sent to each worker thread.

Required Methods

fn new(num_shards: usize) -> Self

Initialize a new allocator with num_shards shards.

num_shards is always nonzero.

fn shards<U, F: Fn(Self::Shard) -> U>(&mut self, vec: &mut Vec<U>, transform: F)

Populates vec with the transformed shards.

The length of the vec must be num_shards in Ealloc::new. The implementation shall shuffle the results returned by this method to avoid centralizing on a single shard.

fn snapshot(&self) -> Snapshot<Self::Raw>

Takes a snapshot of the available entity IDs.

fn allocate(&mut self, hint: Self::AllocHint) -> Self::Raw

Allocate an ID in offline mode.

fn queue_deallocate(&mut self, id: Self::Raw)

Queues the deallocation of an ID.

fn flush(&mut self)

Flushes the queued operations after joining.

fn mark_need_flush(&mut self)

Marks that the allocator needs to flush before executing the next system.

fn flush_if_marked(&mut self)

Flush and reset the mark if mark_need_flush was called since the last flush.

Object Safety

This trait is not object safe.