Trait dynec::entity::ealloc::Recycler

pub trait Recycler<E: Raw>: Default + Extend<E> + Send + 'static {
    type Hint: Default;

    fn len(&self) -> usize;
    fn poll(&mut self, hint: Self::Hint) -> Option<E>;

    fn is_empty(&self) -> bool { ... }
}

A data structure that provides the ability to recycle entity IDs.

Required Associated Types

type Hint: Default

Additional configuration for polling.

Required Methods

fn len(&self) -> usize

Returns the length of this recycler.

fn poll(&mut self, hint: Self::Hint) -> Option<E>

Polls an ID from the recycler based on the given hint.

Provided Methods

fn is_empty(&self) -> bool

Returns whether the recycler is empty.