Trait dynec::entity::ealloc::Shard

pub trait Shard: Send + 'static {
    type Raw: Raw;
    type Hint: Sized;
    type IterAllocatedChunks<'t>: Iterator<Item = Range<Self::Raw>>;

    fn allocate(&mut self, hint: Self::Hint) -> Self::Raw;
    fn iter_allocated_chunks(&self) -> Self::IterAllocatedChunks<'_>;
}

A sharded entity ID allocator.

Each worker thread has mutable access to a shard in each cycle. Between cycles, the shards are shuffled to new worker threads.

Required Associated Types

type Raw: Raw

The raw entity ID type.

type Hint: Sized

The allocation hint for the underlying recycler.

type IterAllocatedChunks<'t>: Iterator<Item = Range<Self::Raw>>

Return value of iter_allocated_chunks.

Required Methods

fn allocate(&mut self, hint: Self::Hint) -> Self::Raw

Allocates an ID from the shard.

fn iter_allocated_chunks(&self) -> Self::IterAllocatedChunks<'_>

Returns an iterator over all allocated ranges during the previous join.

Allocations in the current or other shards after the last flush have no effect on the result of this iterator.

The implementor must return disjoint nonempty chunks in strictly increasing order, i.e. for any two consecutive chunks (a..b), (c..d), a < b <= c < d.