Trait dynec::entity::ealloc::Shard

pub trait Shard: Send + 'static {
    type Raw: Raw;
    type Hint: Sized;

    // Required method
    fn allocate(&mut self, hint: Self::Hint) -> Self::Raw;
}

A sharded entity ID allocator.

Each worker thread has mutable access to a shard in each cycle. Between cycles, the shards are shuffled to new worker threads.

Required Associated Types

type Raw: Raw

The raw entity ID type.

type Hint: Sized

The allocation hint for the underlying recycler.

Required Methods

fn allocate(&mut self, hint: Self::Hint) -> Self::Raw

Allocates an ID from the shard.