Trait dynec::entity::raw::Atomic

pub trait Atomic<E: Raw>: Send + Sync + 'static {
    // Required methods
    fn fetch_add(&self, count: usize) -> E;
    fn load(&self) -> E;
    fn load_mut(&mut self) -> E;
}

An atomic variant of Raw.

Required Methods

fn fetch_add(&self, count: usize) -> E

Equivalent to AtomicUsize::fetch_add(self, count, Ordering::SeqCst)

fn load(&self) -> E

Equivalent to AtomicUsize::load(self, Ordering::SeqCst)

fn load_mut(&mut self) -> E

Equivalent to AtomicUsize::get_mut(self).

This is semantically identical to load, but should be slightly faster because it does not require atomic loading.