Module dynec::scheduler

The scheduler manages the execution of systems, including resource negotiation and dependency constraints.

Modules

builder ğŸ”’

executor ğŸ”’

planner ğŸ”’

state ğŸ”’

topology ğŸ”’

Structs

Order ğŸ”’

PartitionIndex

Uniquely identifies a system::Partition.

ResourceAccess ğŸ”’

Scheduler ğŸ”’

SendSystemIndex

Uniquely identifies a system::Sendable.

UnsendArgs ğŸ”’

UnsendSystemIndex

Uniquely identifies a system::Unsendable.

Enums

Node

Identifies a topological node in the schedule of a cycle.

ResourceType ğŸ”’

WakeupState ğŸ”’