Trait dynec::storage::simple::AnySimpleStorage

pub(crate) trait AnySimpleStorage<A: Archetype>: Send + Sync {
    fn as_any(&self) -> &(dyn Any + Send + Sync);
    fn as_any_mut(&mut self) -> &mut (dyn Any + Send + Sync);
    fn fill_init_simple(&mut self, entity: A::RawEntity, comp_map: &mut Map<A>);
    fn has_finalizer(&self, entity: A::RawEntity) -> bool;
    fn clear_entry(&mut self, entity: A::RawEntity);
    fn referrer_dyn<'t>(&'t mut self) -> Box<dyn Object + 't>;
}

Required Methods

fn as_any(&self) -> &(dyn Any + Send + Sync)

fn as_any_mut(&mut self) -> &mut (dyn Any + Send + Sync)

fn fill_init_simple(&mut self, entity: A::RawEntity, comp_map: &mut Map<A>)

Fills a simple component of with the initial value.

fn has_finalizer(&self, entity: A::RawEntity) -> bool

Returns true if C::IS_FINALIZER. and the component exists for the given entity.

fn clear_entry(&mut self, entity: A::RawEntity)

Clears the component data for an entity if any.

fn referrer_dyn<'t>(&'t mut self) -> Box<dyn Object + 't>

Returns a [referrer::Dyn] implementation that visits all components in this storage.

Implementations

impl<A: Archetype> dyn AnySimpleStorage<A>

pub(crate) fn downcast_ref<C: Simple<A>>(&self) -> &C::Storage

pub(crate) fn downcast_mut<C: Simple<A>>(&mut self) -> &mut C::Storage