Trait dynec::storage::Storage

pub trait Storage: Access + Default + Send + Sync + 'static {
    type Iter<'t>: Iterator<Item = (Self::RawEntity, &'t Self::Comp)> + 't;
    type IterChunks<'t>: Iterator<Item = ChunkRef<'t, Self>> + 't;
    type IterChunksMut<'t>: Iterator<Item = ChunkMut<'t, Self>> + 't;
    type Partition<'u>: Partition<'u, RawEntity = Self::RawEntity, Comp = Self::Comp>
       where Self: 'u;

    // Required methods
    fn get(&self, id: Self::RawEntity) -> Option<&Self::Comp>;
    fn set(
        &mut self,
        id: Self::RawEntity,
        value: Option<Self::Comp>
    ) -> Option<Self::Comp>;
    fn cardinality(&self) -> usize;
    fn iter(&self) -> Self::Iter<'_>;
    fn iter_chunks(&self) -> Self::IterChunks<'_>;
    fn iter_chunks_mut(&mut self) -> Self::IterChunksMut<'_>;
    fn as_partition(&mut self) -> Self::Partition<'_>;
}

A storage for storing component data.

Required Associated Types

type Iter<'t>: Iterator<Item = (Self::RawEntity, &'t Self::Comp)> + 't

Return value of iter.

type IterChunks<'t>: Iterator<Item = ChunkRef<'t, Self>> + 't

Return value of iter_chunks.

type IterChunksMut<'t>: Iterator<Item = ChunkMut<'t, Self>> + 't

Return value of iter_chunks_mut.

type Partition<'u>: Partition<'u, RawEntity = Self::RawEntity, Comp = Self::Comp> where Self: 'u

Return value of as_partition.

Required Methods

fn get(&self, id: Self::RawEntity) -> Option<&Self::Comp>

Gets a shared reference to the component for a specific entity if it is present.

fn set( &mut self, id: Self::RawEntity, value: Option<Self::Comp> ) -> Option<Self::Comp>

Sets or removes the component for a specific entity, returning the original value if it was present.

fn cardinality(&self) -> usize

Returns the number of components that exist in this storage.

fn iter(&self) -> Self::Iter<'_>

Returns an immutable iterator over the storage, ordered by entity index order.

fn iter_chunks(&self) -> Self::IterChunks<'_>

Returns an immutable iterator of slices over the storage, ordered by entity index order.

Each item yielded by the iterator is a tuple of (index, slice), where slice is the slice of components in the chunk, and index is the entity index of slice[0]. slice is always nonempty.

Non-chunked storages should implement this function by returning a chunk for each entity.

fn iter_chunks_mut(&mut self) -> Self::IterChunksMut<'_>

Returns a mutable iterator of slices over the storage, ordered by entity index order.

Each item yielded by the iterator is a tuple of (index, slice), where slice is the slice of components in the chunk, and index is the entity index of slice[0]. slice is always nonempty.

Non-chunked storages should implement this function by returning a chunk for each entity.

fn as_partition(&mut self) -> Self::Partition<'_>

Converts the storage to a Partition that covers the whole storage (similar to slice[..]).

Object Safety

This trait is not object safe.