Module dynec::system

Systems are actions performed every frame to manipulate entities and components.

A system can declare mutable or immutable access to archetype-specific components and global states. dynec will schedule systems such that systems have unique access to mutable resources they request. Furthermore, systems can request relative execution order through Partitions. Systems that use thread-unsafe resources (systems that are not Send) are always executed on the main thread.

Re-exports

pub use access::Isotope as AccessIsotope;

pub use access::Single as AccessSingle;

pub use iter::EntityIterator;

pub use iter::IntoZip;

pub use iter::Try;

pub use iter::Zip;

pub use iter::ZipChunked;

pub use partition::EntityCreationPartition;

pub use partition::Partition;

Modules

access

Access component storages in the world.

iter

Iterate over entities of an archetype.

partition

Partitions enforce execution order of systems. See Partition documentation for more.

spec

Specifies the requirements for a system.

Structs

EntityCreator

Allows creating entities of an archetype.

EntityDeleter

Allows deleting entities of an archetype.

Spec

Describes an instance of system.

Traits

Descriptor

Describes a system.

Sendable

A system requests some resources, stores some states of its own and is runnable with the requested resources.

Unsendable

A variant of Sendable that runs on the main thread only, but allows storing Send states and accessing non-Send + Sync global states.

Type Aliases

ReadIsotopeFull

Provides access to an isotope component in a specific archetype.

ReadIsotopePartial

Provides access to an isotope component in a specific archetype.

ReadSimple

Provides access to a simple component in a specific archetype.

WriteIsotopeFull

Provides access to an isotope component in a specific archetype.

WriteIsotopePartial

Provides access to an isotope component in a specific archetype.

WriteSimple

Provides access to a simple component in a specific archetype.