Module dynec::system

Systems are actions performed every frame to manipulate entities and components.

A system can declare mutable or immutable access to archetype-specific components and global states. dynec will schedule systems such that systems have unique access to mutable resources they request. Furthermore, systems can request relative execution order through Partitions. Systems that use thread-unsafe resources (systems that are not Send) are always executed on the main thread.

Re-exports

pub use partition::EntityCreationPartition;

pub use partition::Partition;

Modules

partition

Partitions enforce execution order of systems. See Partition documentation for more.

spec

Specifies the requirements for a system.

Structs

Spec

Describes an instance of system.

Traits

Accessor

An accessor that can be used in an entity iteration.

ChunkedAccessor

An accessor that can be used in chunked entity iteration.

Descriptor

Describes a system.

EntityCreator

Allows creating entities of an archetype.

EntityDeleter

Allows deleting entities of an archetype.

EntityIterator

Allows iterating all entities of an archetype.

Mut

Generalizes WriteSimple, WriteIsotope and their split storages.

Read

Generalizes ReadSimple and ReadIsotope for a specific discriminant (through ReadIsotope::split).

ReadChunk

Extends Read with chunk reading ability for storages that support chunked access.

ReadIsotope

Provides access to an isotope component in a specific archetype.

ReadSimple

Provides access to a simple component in a specific archetype.

Sendable

A system requests some resources, stores some states of its own and is runnable with the requested resources.

Unsendable

A variant of Sendable that runs on the main thread only, but allows storing Send states and accessing non-Send + Sync global states.

Write

Generalizes WriteSimple and WriteIsotope for a specific discriminant (through WriteIsotope::split_isotopes).

WriteChunk

Extends Write with chunk writing ability for storages that support chunked access.

WriteIsotope

Provides access to an isotope component in a specific archetype.

WriteSimple

Provides access to a simple component in a specific archetype.