Trait dynec::system::accessor::Accessor

pub unsafe trait Accessor<A: Archetype> {
    type Entity<'t>: 't
       where Self: 't;

    // Required method
    unsafe fn entity<'this, 'e, 'ret>(
        this: &'this mut Self,
        id: TempRef<'e, A>
    ) -> Self::Entity<'ret>
       where Self: 'ret;
}

An accessor that can be used in an entity iteration.

Safety

Implementors must ensure that entity is deterministic and one-to-one.

Multiplexing implementors (such as tuples or composite accessors) preserve this invariant automatically since they are just destructuring to independent storages. Storage delegations preserve this invariant automatically since Storage::get_mut has the same safety invariants (see Storage  Safety).

Required Associated Types

type Entity<'t>: 't where Self: 't

Return value of entity.

Required Methods

unsafe fn entity<'this, 'e, 'ret>( this: &'this mut Self, id: TempRef<'e, A> ) -> Self::Entity<'ret>where Self: 'ret,

Accesses this storage for a specific entity.

Safety

The lifetime of the return value is arbitrarily defined by the caller. This effectively disables the borrow checker for return values. The caller must ensure that return values do not outlive self, and the function result is dropped before it is called again with the same id.