Function dynec::system::entity::entity_iterator

pub fn entity_iterator<A: Archetype>(
    ealloc: Snapshot<A::RawEntity>
) -> impl EntityIterator<A>

Constructs an instance of EntityIterator that reads from the given allocator.

Although this function accepts an allocator shard, it actually reads the global buffer shared between shards, which is independent of the changes in the current shard. Hence, the iterator describe the state after the previous tick completes, which does not include newly initialized entities and includes those queued for deletion. This behavior is reasonable, because newly initialized entities should not be accessed at all, and those queued for deletion may have a finalizer or be given a finalizer when running later systems, so those queued for deletion are still included.

This function is typically called from the code generated by #[system].