Trait dynec::system::iter::IntoZip

pub trait IntoZip<A: Archetype> {
    type IntoZip: Zip<A>;

    // Required method
    fn into_zip(self) -> Self::IntoZip;
}

Values that can be used as a Zip in EntityIterator, similar to IntoIterator for iterators.

This trait is intended to map storages to components of a single entity, so it is implemented by:

&ReadSimple and &mut WriteSimple

Shared/mutable references to split isotope accessors

Any of the above wrapped with Try for optional components.

Tuples of Zip implementors, including other tuples.

Structs of Zip fields that use the Zip derive macro.

The default configuration only implements for tuples of up to 4 elements. To use larger tuples at the cost of slower compile time, use the feature "tuple-impl-{n}-zip", where {n} is 8, 16, 24 or 32.

Required Associated Types

type IntoZip: Zip<A>

The Zip type that this is converted into.

Required Methods

fn into_zip(self) -> Self::IntoZip

Converts into a Zip object.