Trait dynec::system::iter::Zip

pub trait Zip<A: Archetype>: Sized {
    type Item;

    // Required methods
    fn split(&mut self, offset: A::RawEntity) -> Self;
    fn get<E: Ref<Archetype = A>>(self, entity: E) -> Self::Item;
}

Multiple single accessors zipped together, to be used with EntityIterator::entities_with.

All accessors must target the same archetype A.

See IntoZip for what values can be passed for Zip.

Required Associated Types

type Item

The type of values available for a single entity.

Required Methods

fn split(&mut self, offset: A::RawEntity) -> Self

Vertically splits each underlying storage vertically (by entities) at offset.

fn get<E: Ref<Archetype = A>>(self, entity: E) -> Self::Item

Returns the requested components for the specified entity.

Object Safety

This trait is not object safe.