Trait dynec::system::iter::ZipChunked

pub trait ZipChunked<A: Archetype>: Zip<A> {
    type Chunk;

    // Required methods
    fn get_chunk(self, chunk: TempRefChunk<'_, A>) -> Self::Chunk;
    fn chunk_to_entities(chunk: Self::Chunk) -> impl Iterator<Item = Self::Item>;
}

Zip accessors with the additional condition that all underlying storages support chunked access, to be used with EntityIterator::chunks_with.

Required Associated Types

type Chunk

The type of values available for a single chunk.

Required Methods

fn get_chunk(self, chunk: TempRefChunk<'_, A>) -> Self::Chunk

Returns the requested components as chunks for the specified entities.

fn chunk_to_entities(chunk: Self::Chunk) -> impl Iterator<Item = Self::Item>

Converts a chunk into an iterator of items.

Object Safety

This trait is not object safe.