Trait dynec::system::ChunkedAccessor

pub unsafe trait ChunkedAccessor<A: Archetype> {
    type Chunk<'t>: 't
    where
        Self: 't;

    unsafe fn chunk<'this, 'e, 'ret>(
        this: &'this mut Self,
        chunk: TempRefChunk<'e, A>
    ) -> Self::Chunk<'ret>;
}

An accessor that can be used in chunked entity iteration.

Safety

Implementors must ensure that chunk is deterministic, and non-overlapping entity chunks return non-overlapping values. This is equivalent to (and should delegate to) [storage::Chunked::get_chunk]/[storage::Chunked::get_chunk_mut].

Multiplexing implementors (such as tuples or composite accessors) preserve this invariant automatically since they are just destructuring to independent storages. Storage delegations preserve this invariant automatically since [storage::Chunked::get_chunk_mut] has the same safety invariants (see Chunked  Safety).

Required Associated Types

type Chunk<'t>: 't
where
Self: 't

Return value of chunk.

Required Methods

unsafe fn chunk<'this, 'e, 'ret>(
this: &'this mut Self,
chunk: TempRefChunk<'e, A>
) -> Self::Chunk<'ret>

Accesses this storage for a specific chunk of entities.

Safety

The lifetime of the return value is arbitrarily defined by the caller. This effectively disables the borrow checker for return values. The caller must ensure that return values do not outlive self, and the function result is dropped before it is called again with an overlapping chunk.