Trait dynec::system::Mut

pub trait Mut<A: Archetype, C: 'static> {
    type IterMut<'t>: Iterator<Item = (TempRef<'t, A>, &'t mut C)>
    where
        Self: 't;
    type SplitEntitiesAt<'u>: Mut<A, C> + 'u
    where
        Self: 'u;

    fn try_get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> Option<&mut C>;
    fn iter_mut(&mut self) -> Self::IterMut<'_>;
    fn split_entities_at<E: Ref<Archetype = A>>(
        &mut self,
        entity: E
    ) -> (Self::SplitEntitiesAt<'_>, Self::SplitEntitiesAt<'_>);
}

Generalizes WriteSimple, WriteIsotope and their split storages.

Only supports mutable access to an existing component, but does not support adding or removing components since only the storage values but not the storage structure can be borrowed mutably.

Required Associated Types

type IterMut<'t>: Iterator<Item = (TempRef<'t, A>, &'t mut C)>
where
Self: 't

Return value of iter_mut.

type SplitEntitiesAt<'u>: Mut<A, C> + 'u
where
Self: 'u

Return value of split_entities_at.

Required Methods

fn try_get_mut<E: Ref<Archetype = A>>(&mut self, entity: E) -> Option<&mut C>

Returns a mutable reference to the component for the specified entity, or None if the component is not present in the entity.

Note that this method returns Option<&mut C>, not &mut Option<C>. This means setting the Option itself to Some/None will not modify any stored value. Use Write::set to add/remove a component.

fn iter_mut(&mut self) -> Self::IterMut<'_>

Iterates over mutable references to all initialized components in this storage.

fn split_entities_at<E: Ref<Archetype = A>>(
&mut self,
entity: E
) -> (Self::SplitEntitiesAt<'_>, Self::SplitEntitiesAt<'_>)

Partitions the accessor into two disjoint halves of entities.

This method is not required for Read because shared references can be reused directly.