Trait dynec::system::Read

pub trait Read<A: Archetype, C: 'static> {
    type Iter<'t>: Iterator<Item = (TempRef<'t, A>, &'t C)>
    where
        Self: 't;
    type DuplicateImmut<'t>: Read<A, C> + 't
    where
        Self: 't;

    fn try_get<E: Ref<Archetype = A>>(&self, entity: E) -> Option<&C>;
    fn iter(&self) -> Self::Iter<'_>;
    fn duplicate_immut(
        &self
    ) -> (Self::DuplicateImmut<'_>, Self::DuplicateImmut<'_>);

    fn get<E: Ref<Archetype = A>>(&self, entity: E) -> &C
    where
        C: Must<A>,
    { ... }
    fn access(&self) -> MustRead<A, C, &Self>
    where
        C: Must<A>,
    { ... }
    fn try_access(&self) -> TryRead<A, C, &Self> { ... }
}

Generalizes ReadSimple and ReadIsotope for a specific discriminant (through ReadIsotope::split).

Required Associated Types

type Iter<'t>: Iterator<Item = (TempRef<'t, A>, &'t C)>
where
Self: 't

Return value of iter.

type DuplicateImmut<'t>: Read<A, C> + 't
where
Self: 't

Return value of duplicate_immut.

Required Methods

fn try_get<E: Ref<Archetype = A>>(&self, entity: E) -> Option<&C>

Returns an immutable reference to the component for the specified entity, or None if the component is not present in the entity.

fn iter(&self) -> Self::Iter<'_>

Iterates over all initialized components in this storage.

fn duplicate_immut(
&self
) -> (Self::DuplicateImmut<'_>, Self::DuplicateImmut<'_>)

Duplicates the current reader, producing two new values that can only access the storage immutably.

Provided Methods

fn get<E: Ref<Archetype = A>>(&self, entity: E) -> &Cwhere
C: Must<A>,

Returns an immutable reference to the component for the specified entity.

Panics

This method panics if the entity is not fully initialized yet. This happens when an entity is newly created and the cycle hasn’t joined yet.

fn access(&self) -> MustRead<A, C, &Self>where
C: Must<A>,

Returns an Accessor implementor that yields &C for each entity.

fn try_access(&self) -> TryRead<A, C, &Self>

Returns an Accessor implementor that yields Option<&C> for each entity.