Trait dynec::system::ReadChunk

pub trait ReadChunk<A: Archetype, C: 'static> {
    fn get_chunk(&self, chunk: TempRefChunk<'_, A>) -> &[C] 
    where
        C: Must<A>;
}

Extends Read with chunk reading ability for storages that support chunked access.

Required Methods

fn get_chunk(&self, chunk: TempRefChunk<'_, A>) -> &[C] where
C: Must<A>,

Returns the chunk of components as a slice.

Panics

This method panics if any component in the chunk is missing. In general, users should not get an entity::TempRefChunk that includes an uninitialized entity, so panic is basically impossible if comp::Must was implemented correctly.