Trait dynec::util::UnsafeEqOrd

pub unsafe trait UnsafeEqOrd: Eq + Ord { }

Same as Eq and Ord, but with a stronger guarantee.

Safety

Undefined behavior may occur if the invariants of Eq and Ord are not fully satisfied.