Module dynec::world

The world stores the states of the game.

Re-exports

pub use storage::Storage;

pub use tracer::Tracer;

Modules

accessor 🔒

builder 🔒

global 🔒

offline

Operations queued to be executed after the cycle joins.

scheduler 🔒

storage

A storage is the data structure where components of the same type for all entities are stored.

tracer

Exposes testing, profiling and tracing capabilities.

typed 🔒

Structs

Builder

This type is used to build a world. No more systems can be scheduled after the builder is built.

Components

Stores the component states in a world.

PartitionIndex

Uniquely identifies a system::Partition.

SendSystemIndex

Uniquely identifies a system::Sendable.

SyncGlobals

Stores the thread-safe global states in a world.

UnsendSystemIndex

Uniquely identifies a system::Unsendable.

UnsyncGlobals

Stores the thread-unsafe global states in a world.

World

The data structure that stores all states in the game.

Enums

DeleteResult 🔒

ScheduleNode

Identifies a topological node in the schedule of a cycle.

Traits

Bundle

A bundle encapsulates the systems and resources for a specific feature. This can be used by library crates to expose their features as a single API.

Functions

flag_delete_entity 🔒

Flags an entity for deletion, and deletes it immediately if there are no finalizers.

init_entity 🔒

Initializes an entity after allocation.

new

Creates a dynec world from bundles.

new_unthreaded

Creates a dynec world from bundles with threading disabled.

new_with_concurrency

Creates a dynec world from bundles and specify the number of worker threads (not counting the main thread, which only executes thread-local tasks).

search_references 🔒

try_real_delete_entity 🔒

Deletes an entity immediately if there are no finalizers.