Trait dynec::world::storage::Storage

pub unsafe trait Storage: Default + Send + Sync + 'static {
    type RawEntity: Raw;
    type Comp;
    type Iter<'t>: Iterator<Item = (Self::RawEntity, &'t Self::Comp)> + 't;
    type IterChunks<'t>: Iterator<Item = ChunkRef<'t, Self>> + 't;
    type IterMut<'t>: Iterator<Item = (Self::RawEntity, &'t mut Self::Comp)> + 't;
    type IterChunksMut<'t>: Iterator<Item = ChunkMut<'t, Self>> + 't;

    fn get(&self, id: Self::RawEntity) -> Option<&Self::Comp>;
    fn get_mut(&mut self, id: Self::RawEntity) -> Option<&mut Self::Comp>;
    fn set(
        &mut self,
        id: Self::RawEntity,
        value: Option<Self::Comp>
    ) -> Option<Self::Comp>;
    fn cardinality(&self) -> usize;
    fn iter(&self) -> Self::Iter<'_>;
    fn iter_chunks(&self) -> Self::IterChunks<'_>;
    fn iter_mut(&mut self) -> Self::IterMut<'_>;
    fn iter_chunks_mut(&mut self) -> Self::IterChunksMut<'_>;
}

A storage for storing component data.

Safety

Implementors of this trait must ensure that get and get_mut are consistent and injective. In other words, for any a: Self::RawEntity, get(a) and get_mut(a) return the same value (only differing by mutability), and for any b: Self::RawEntity where a != b, get(a) must not alias get(b).

This implies the implementation is not safe if Eq and Ord are incorrectly implemented for Self::RawEntity, which is why entity::Raw is an unsafe trait that strictly requires complete equivalence and ordering. (Imagine if RawEntity is f64, and a and b are both f64::NAN; then a != b but get_mut(a) would still alias get_mut(b))

Required Associated Types

type RawEntity: Raw

The type of entity ID used for identification.

type Comp

The component type stored.

type Iter<'t>: Iterator<Item = (Self::RawEntity, &'t Self::Comp)> + 't

Return value of iter.

type IterChunks<'t>: Iterator<Item = ChunkRef<'t, Self>> + 't

Return value of iter_chunk.

type IterMut<'t>: Iterator<Item = (Self::RawEntity, &'t mut Self::Comp)> + 't

Return value of iter_mut.

type IterChunksMut<'t>: Iterator<Item = ChunkMut<'t, Self>> + 't

Return value of iter_chunk_mut.

Required Methods

fn get(&self, id: Self::RawEntity) -> Option<&Self::Comp>

Gets a shared reference to the component for a specific entity if it is present.

fn get_mut(&mut self, id: Self::RawEntity) -> Option<&mut Self::Comp>

Gets a mutable reference to the component for a specific entity if it is present.

fn set(
&mut self,
id: Self::RawEntity,
value: Option<Self::Comp>
) -> Option<Self::Comp>

Sets or removes the component for a specific entity, returning the original value if it was present.

fn cardinality(&self) -> usize

Returns the number of components that exist in this storage.

fn iter(&self) -> Self::Iter<'_>

Returns an immutable iterator over the storage, ordered by entity index order.

fn iter_chunks(&self) -> Self::IterChunks<'_>

Returns an immutable iterator of slices over the storage, ordered by entity index order.

Each item yielded by the iterator is a tuple of (index, slice), where slice is the slice of components in the chunk, and index is the entity index of slice[0]. slice is always nonempty.

fn iter_mut(&mut self) -> Self::IterMut<'_>

Returns a mutable iterator over the storage, ordered by entity index order.

fn iter_chunks_mut(&mut self) -> Self::IterChunksMut<'_>

Returns a mutable iterator of slices over the storage, ordered by entity index order.

Each item yielded by the iterator is a tuple of (index, slice), where slice is the slice of components in the chunk, and index is the entity index of slice[0]. slice is always nonempty.