Macro dynec::world::tracer::define_tracer

macro_rules! define_tracer {
    (
        $(
            $(#[$meta:meta])*
            fn $name:ident(
                &self
                $(,$logged_ident:ident: $logged_ty:ty $({@LOG_WITH = $log_with:expr})?)*
                $(; @NOLOG $($nolog_ident:ident: $nolog_ty:ty),*)?
                $(,)?
            );
        )*
    ) => { ... };
    (@LOG_EXPR $ident:ident) => { ... };
    (@LOG_EXPR $ident:ident @LOG_WITH = $closure:expr) => { ... };
}

Defines the Tracer trait and implements the Log and Aggregate types.

All tracer method parameters must be either Copy or a mutable reference (or immutable reference, which is Copy).

Use the {@LOG_WITH = transformer} syntax to transform an argument for log printing such that transformer(argument) returns a fmt::Debug type.