Struct dynec::world::tracer::Aggregate

pub struct Aggregate<T>(pub T);

Groups multiple tracers into a tuple and dispatches each call to them in serial.

Tuple Fields
0: T

A tuple of child tracers to execute in serial.