Trait dynec::world::tracer::Tracer

pub trait Tracer: Sync {
    fn start_cycle(&self) { ... }
    fn end_cycle(&self) { ... }
    fn steal_return_complete(&self, thread: Thread) { ... }
    fn steal_return_pending(&self, thread: Thread) { ... }
    fn mark_runnable(&self, node: ScheduleNode) { ... }
    fn unmark_runnable(&self, node: ScheduleNode) { ... }
    fn start_run_sendable(
        &self, 
        thread: Thread, 
        node: ScheduleNode, 
        debug_name: &str, 
        system: &mut dyn Sendable
    ) { ... }
    fn end_run_sendable(
        &self, 
        thread: Thread, 
        node: ScheduleNode, 
        debug_name: &str, 
        system: &mut dyn Sendable
    ) { ... }
    fn start_run_unsendable(
        &self, 
        thread: Thread, 
        node: ScheduleNode, 
        debug_name: &str, 
        system: &mut dyn Unsendable
    ) { ... }
    fn end_run_unsendable(
        &self, 
        thread: Thread, 
        node: ScheduleNode, 
        debug_name: &str, 
        system: &mut dyn Unsendable
    ) { ... }
    fn partition(&self, node: ScheduleNode, partition: &dyn Partition) { ... }
}

A tracer used for recording the events throughout an execution cycle.

Can be used for profiling and testing.

Provided Methods

fn start_cycle(&self)

A cycle starts.

fn end_cycle(&self)

A cycle ends.

fn steal_return_complete(&self, thread: Thread)

A thread tries to steal a task, but all tasks have started.

fn steal_return_pending(&self, thread: Thread)

A thread tries to steal a task, but no tasks are in the runnable pool.

fn mark_runnable(&self, node: ScheduleNode)

A node is marked as runnable because all blockers have been removed.

fn unmark_runnable(&self, node: ScheduleNode)

A node is unmarked as runnable because an exclusive node has been stolen.

fn start_run_sendable(
&self,
thread: Thread,
node: ScheduleNode,
debug_name: &str,
system: &mut dyn Sendable
)

A thread-safe system starts running.

fn end_run_sendable(
&self,
thread: Thread,
node: ScheduleNode,
debug_name: &str,
system: &mut dyn Sendable
)

A thread-safe system stops running.

fn start_run_unsendable(
&self,
thread: Thread,
node: ScheduleNode,
debug_name: &str,
system: &mut dyn Unsendable
)

A thread-unsafe system starts running.

fn end_run_unsendable(
&self,
thread: Thread,
node: ScheduleNode,
debug_name: &str,
system: &mut dyn Unsendable
)

A thread-unsafe system stops running.

fn partition(&self, node: ScheduleNode, partition: &dyn Partition)

A partition completes.